+++ include/mathalea_exercices_3e.js
@@ -1,1755 +1,5 @@
-/**
- * Trace 5 droites et demande l'expression de la fonction affine ou linéaire correspondante
- * @Auteur Jean-Claude Lhote
- * Référence : 3F21
- */
-function Lecture_expression_fonctions_affines() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Déterminer une fonction affine";
-	this.consigne = "Donner l'expression des fonctions représentées";
-	this.nb_questions = 1;
-	this.nb_questions_modifiable = false;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	sortie_html ? this.spacing = 2 : this.spacing = 1;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-	this.sup = 1;
-	this.sup2 = 3
-	this.lineaire = false;
-	this.liste_packages = 'tkz-euclide';
 
-
-	this.nouvelle_version = function (numero_de_l_exercice) { // numero_de_l_exercice est 0 pour l'exercice 1
-		let k = Math.pow(2, parseInt(this.sup) - 1);
-		let nb_droites = parseInt(this.sup2)
-		this.liste_questions = [];
-		this.liste_corrections = [];
-		this.contenu = ''; // Liste de questions
-		this.contenu_correction = ''; // Liste de questions corrigées
-		let h = Math.round(window.innerHeight * 0.7) //pour déterminer la hauteur du div 
-		let liste_droites = [];
-		let OrdX0;
-		let Pente = [];
-		if (!this.lineaire) {
-			Pente.push(randint(-2 * k, 2 * k));
-			Pente.push(randint(-2 * k, 2 * k, [Pente[0]]));
-			Pente.push(randint(-2 * k, 2 * k, [Pente[0], Pente[1]]));
-			Pente.push(randint(-2 * k, 2 * k, [Pente[0], Pente[1], Pente[2]]));
-			Pente.push(randint(-2 * k, 2 * k, [Pente[0], Pente[1], Pente[2], Pente[3]]));
-		}
-		else {
-			Pente.push(randint(-3 * k, 3 * k, [0]));
-			Pente.push(randint(-3 * k, 3 * k, [Pente[0], 0]));
-			Pente.push(randint(-3 * k, 3 * k, [Pente[0], Pente[1], 0]));
-			Pente.push(randint(-3 * k, 3 * k, [Pente[0], Pente[1], Pente[2], 0]));
-			Pente.push(randint(-3 * k, 3 * k, [Pente[0], Pente[1], Pente[2], Pente[3], 0]));
-		}
-
-		for (let i = 0; i < 5; i++) {
-			if (this.lineaire) OrdX0 = 0;
-			else OrdX0 = randint(-1 + Pente[i] / k, 1 + Pente[i] / k)
-			liste_droites.push([OrdX0, Pente[i] / k])
-		}
-
-		if (sortie_html) {
-			const mon_svg = SVG().viewbox(0, 0, 500, 500).size('100%', '100%')
-			SVG_repere(mon_svg, -5, 5, -5, 5, k, k, 500, 500, true);
-			SVG_Tracer_droite(mon_svg, 500, 500, -5, 5, -5, 5, liste_droites[0][0], liste_droites[0][1], 'blue', 'd1');
-			if (nb_droites > 1) SVG_Tracer_droite(mon_svg, 500, 500, -5, 5, -5, 5, liste_droites[1][0], liste_droites[1][1], 'red', 'd2');
-			if (nb_droites > 2) SVG_Tracer_droite(mon_svg, 500, 500, -5, 5, -5, 5, liste_droites[2][0], liste_droites[2][1], 'green', 'd3');
-			if (nb_droites > 3) SVG_Tracer_droite(mon_svg, 500, 500, -5, 5, -5, 5, liste_droites[3][0], liste_droites[3][1], 'brown', 'd4');
-			if (nb_droites > 4) SVG_Tracer_droite(mon_svg, 500, 500, -5, 5, -5, 5, liste_droites[4][0], liste_droites[4][1], 'purple', 'd5');
-			this.consigne = `<div style="width: 50%; display : table ">${mon_svg.svg()}</div>`;
-
-
-
-		}
-		else { //sortie Latex 
-			let texte = `\\begin{tikzpicture}`;
-			texte += Latex_repere(-5, 5, -5, 5, k, k, true);
-			texte += Latex_Tracer_droite(-5, 5, -5, 5, liste_droites[0][0], liste_droites[0][1], 'blue', 'd_1');
-			if (nb_droites > 1) texte += Latex_Tracer_droite(-5, 5, -5, 5, liste_droites[1][0], liste_droites[1][1], 'red', 'd_2');
-			if (nb_droites > 2) texte += Latex_Tracer_droite(-5, 5, -5, 5, liste_droites[2][0], liste_droites[2][1], 'green', 'd_3');
-			if (nb_droites > 3) texte += Latex_Tracer_droite(-5, 5, -5, 5, liste_droites[3][0], liste_droites[3][1], 'brown', 'd_4');
-			if (nb_droites > 4) texte += Latex_Tracer_droite(-5, 5, -5, 5, liste_droites[4][0], liste_droites[4][1], 'purple', 'd_5');
-			texte += `\n\t \\end{tikzpicture}`;
-			this.liste_questions.push(texte);
-		}
-		for (i = 0; i < nb_droites; i++) {
-			this.liste_questions.push(`Déterminer l'expression de la fonction $f_${i + 1}$ représentée par la droite $d_${i + 1}$.`)
-			if (this.lineaire || liste_droites[i][0] == 0) this.liste_corrections.push(`La droite $d_${i + 1}$ passe par l'origine et son coefficient directeur est $${tex_nombre(liste_droites[i][1])}$.<br>Elle représente la fonction linéaire $f_${i + 1}(x)=${reduire_ax_plus_b(liste_droites[i][1], 0)}$.`)
-			else this.liste_corrections.push(`La droite $d_${i + 1}$ passe par le point de coordonnées $(0;${liste_droites[i][0]})$ et son coefficient directeur est $${tex_nombre(liste_droites[i][1])}$.<br>Elle représente la fonction affine $f_${i + 1}(x)=${reduire_ax_plus_b(liste_droites[i][1], liste_droites[i][0])}$.`)
-
-		}
-
-		liste_de_question_to_contenu_sans_numero(this);
-		if (!this.lineaire) this.contenu_correction = `Il s’agit de fonctions affines, elles sont donc de la forme $f(x)=ax+b$, $b$ étant l’ordonnée à l’origine et $a$ la pente de la droite.\n` + this.contenu_correction;
-		else this.contenu_correction = `Il s’agit de fonctions linéaires, elles sont donc de la forme $f(x)=ax$, $a$ étant la ` + katex_Popup2(numero_de_l_exercice, 1, `pente`, `pente d'une droite`, `La pente (le a de y=ax ou y=ax+b) d'une droite donne le taux d'accroissement de y par rapport à x : lorsque x augmente de 1, alors y augmente de a.`) + ` de la droite.\n` + this.contenu_correction;
-	}
-	this.besoin_formulaire_numerique = ['Niveau de difficulté', 3, "1 : Coefficient directeur entier\n2 : Coefficient directeur 'en demis'\n3 : Coefficient directeur 'en quarts'"];
-	this.besoin_formulaire2_numerique = ['Nombre de droites (1 à 5)', 5];
-}
-
-/**
-* Réduire des expressions lorsque c'est possible
-*
-* @Auteur Rémi Angot
-* 3L11-2
-*/
-function Reduction_si_possible() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Réduire une expression";
-	this.consigne = "Réduire les expressions suivantes, si cela est possible.";
-	this.nb_questions = 5;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-
-		let type_de_questions_disponibles = ['ax+b', 'ax+bx', 'ax+bx2', 'ax*b', 'b*ax', 'ax+b+cx+d', 'b+ax+d+cx', 'ax+b+x'];
-		let liste_type_de_questions = combinaison_listes(type_de_questions_disponibles, this.nb_questions); // Tous les types de questions sont posées mais l'ordre diffère à chaque "cycle"
-		for (let i = 0, texte, texte_corr, a, b, c, d, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			a = randint(-11, 11, 0);
-			b = randint(-11, 11, [0, a]);
-			c = randint(-11, 11, [0]);
-			d = randint(-11, 11, 0)
-			switch (liste_type_de_questions[i]) {
-				case 'ax+b':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${a}*x+(${b})`)}$`;
-					texte_corr = texte
-					break;
-				case 'ax+bx':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${a}*x+(${b}*x)`)}$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${a}*x+(${b}*x)`)}=${printlatex(`${a + b}x`)}$`;
-					break;
-				case 'ax+bx2':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${a}*x+(${b}*x^2)`)}$`;
-					texte_corr = texte
-					break;
-				case 'ax*b':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${a}*x`)}\\times ${ecriture_parenthese_si_negatif(b)}$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${a}*x`)}\\times ${ecriture_parenthese_si_negatif(b)}=${printlatex(`${a * b}*x`)}$`;
-					break;
-				case 'b*ax':
-					a = randint(1, 11);
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${b}\\times ${printlatex(`${a}*x`)}$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=${b}\\times ${printlatex(`${a}*x`)}=${printlatex(`${b * a}*x`)}$`;
-					break;
-				case 'ax+b+cx+d':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${a}*x+(${b})+(${c})*x+(${d})`)}$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${a}*x+(${b})+(${c})*x+(${d})`)}`;
-					if (b + d == 0) {
-						if (a + c == 0) {
-							texte_corr += `=0$`
-						} else {
-							texte_corr += `=${printlatex(`${a + c}*x`)}$`
-						}
-					} else {
-						if (a + c == 0) {
-							texte_corr += `=${b + d}$`
-						} else {
-							texte_corr += `=${printlatex(`${a + c}*x+(${b + d})`)}$`
-						}
-					}
-					break;
-				case 'b+ax+d+cx':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${b}+(${a})*x+(${d})+(${c})*x`)}$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${b}+(${a})*x+(${d})+(${c})*x`)}`;
-					if (b + d == 0) {
-						if (a + c == 0) {
-							texte_corr += `=0$`
-						} else {
-							texte_corr += `=${printlatex(`${a + c}*x`)}$`
-						}
-					} else {
-						if (a + c == 0) {
-							texte_corr += `=${b + d}$`
-						} else {
-							texte_corr += `=${printlatex(`${a + c}*x+(${b + d})`)}$`
-						}
-					}
-					break;
-				case 'ax+b+x':
-					a = randint(-11, 11, [0, -1])
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${a}*x+(${b})+x`)}$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${a}*x+(${b})+x`)}=${printlatex(`${a + 1}*x+(${b})`)}$`
-					break;
-
-			}
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++;
-		}
-		liste_de_question_to_contenu(this);
-	}
-}
-
-
-/**
-* Utiliser la simple ou la double distributivité et réduire l'expression
-*
-* @Auteur Rémi Angot
-* 3L11-3
-*/
-function Distributivite_simple_double_reduction() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Utiliser la distributivité (simple ou double) et réduire";
-	this.consigne = "Développer et réduire les expressions suivantes.";
-	this.nb_questions = 5;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-
-		let type_de_questions_disponibles = ['cx+e(ax+b)', 'ex+(ax+b)(cx+d)', 'e+(ax+b)(cx+d)', 'e-(ax+b)(cx+d)', '(ax*b)(cx+d)', 'e(ax+b)-(d+cx)'];
-		let liste_type_de_questions = combinaison_listes(type_de_questions_disponibles, this.nb_questions); // Tous les types de questions sont posées mais l'ordre diffère à chaque "cycle"
-		for (let i = 0, texte, texte_corr, a, b, c, d, e, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			a = randint(-11, 11, 0);
-			b = randint(-11, 11, 0);
-			c = randint(-11, 11, 0);
-			d = randint(-11, 11, 0);
-			e = randint(-11, 11, 0);
-			switch (liste_type_de_questions[i]) {
-				case 'cx+e(ax+b)':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${c}*x+(${e})*(${a}*x+(${b}))`)}$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${c}*x+(${e})*(${a}*x+(${b}))`)}$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`${c}*x+(${e * a})*x+(${e * b})`)}$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`${c + e * a}*x+(${e * b})`)}$`;
-					break;
-				case 'ex+(ax+b)(cx+d)':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${e}*x+(${a}*x+(${b}))*(${c}x+(${d}))`)}$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${e}*x+(${a}*x+(${b}))*(${c}x+(${d}))`)}$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`${e}*x+(${a * c})*x^2+(${a * d})*x+(${b * c})*x+(${b * d})`)}$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`${a * c}*x^2+(${e + b * c + a * d})*x+(${b * d})`)}$`;
-					break;
-				case 'e+(ax+b)(cx+d)':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${e}+(${a}*x+(${b}))*(${c}x+(${d}))`)}$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${e}+(${a}*x+(${b}))*(${c}x+(${d}))`)}$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`${e}+(${a * c})*x^2+(${a * d})*x+(${b * c})*x+(${b * d})`)}$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`${a * c}*x^2+(${b * c + a * d})*x+(${e + b * d})`)}$`;
-					break;
-				case 'e-(ax+b)(cx+d)':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${e}-${printlatex(`(${a}*x+(${b}))*(${c}x+(${d}))`)}$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=${e}-${printlatex(`(${a}*x+(${b}))*(${c}x+(${d}))`)}$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${e}-(${printlatex(`(${a * c})*x^2+(${a * d})*x+(${b * c})*x+(${b * d})`)})$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`${e}+(${-1 * a * c})*x^2+(${-1 * a * d})*x+(${-1 * b * c})*x+(${-1 * b * d})`)}$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`${-1 * a * c}*x^2+(${-1 * b * c - a * d})*x+(${e - b * d})`)}$`;
-					break;
-				case '(ax*b)(cx+d)':
-					a = randint(-3, 3, [0]);
-					b = randint(2, 3);
-					texte = `$${lettre_depuis_chiffre(i + 1)}=(${printlatex(`${a}*x`)}\\times${b})(${printlatex(`${c}*x+(${d})`)})$`;
-					texte_corr = `$${lettre_depuis_chiffre(i + 1)}=(${printlatex(`${a}*x`)}\\times${b})(${printlatex(`${c}*x+(${d})`)})$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`${a * b}*x`)}\\times(${printlatex(`${c}*x+(${d})`)})$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`${a * b * c}*x^2+(${a * b * d})*x`)}$`;
-					break;
-				case 'e(ax+b)-(d+cx)':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${e}(${printlatex(`${a}*x+(${b})`)})-(${printlatex(`${d}+(${c})*x`)})$`;
-					texte_corr = texte;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`(${e * a})*x+(${e * b})`)}-(${printlatex(`${d}+(${c})*x`)})$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`(${e * a})*x+(${e * b})+(${-d})+(${-c})*x`)}$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${printlatex(`(${e * a - c})*x+(${e * b - d})`)}$`;
-					break;
-
-
-			}
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++;
-		}
-		liste_de_question_to_contenu(this);
-	}
-}
-
-/**
-* Utiliser la simple ou la double distributivité et réduire l'expression
-*
-* @Auteur Rémi Angot
-* 3L11-4
-*/
-function Factoriser_par_nombre_ou_x() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Factoriser une expression";
-	this.consigne = "Factoriser les expressions suivantes.";
-	this.nb_questions = 8;
-	this.nb_cols = 2;
-	this.nb_cols_corr = 2;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-
-		let type_de_questions_disponibles = ['ka+nkb', '-ka+nkb', 'nka+mkb', 'nka-mkb', 'nkx+mkx2', 'nkx-mkx2', 'nx2+x', 'nx2+mx'];
-		let liste_type_de_questions = combinaison_listes(type_de_questions_disponibles, this.nb_questions); // Tous les types de questions sont posées mais l'ordre diffère à chaque "cycle"
-		for (let i = 0, texte, texte_corr, n, m, couplenm, k, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			k = choice([2, 3, 5, 7, 11]);
-			couplenm = choice([[2, 3], [3, 4], [2, 5], [3, 5], [4, 5], [5, 6], [2, 7], [3, 7], [4, 7], [5, 7], [6, 7], [3, 8], [5, 8], [7, 8], [2, 9], [4, 9], [5, 9], [7, 9], [8, 9], [3, 10], [7, 10], [9, 10]]); // n et m sont premiers entre eux
-			n = couplenm[0];
-			m = couplenm[1];
-			switch (liste_type_de_questions[i]) {
-				case 'ka+nkb':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${k}*a+(${n * k})*b`)}$`;
-					texte_corr = texte;
-					if (n > 0) {
-						texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}a+${k}\\times${n}b$`;
-					} else {
-						texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}a-${k}\\times${abs(n)}b$`;
-					}
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}(${printlatex(`a+(${n})*b`)})$`;
-					break;
-				case '-ka+nkb':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${-k}*a+(${n * k})*b`)}$`;
-					texte_corr = texte;
-					if (n > 0) {
-						texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${-k}a+${k}\\times${n}b$`;
-						texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}(${printlatex(`-a+${n}*b`)})$`;
-					} else {
-						texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${-k}a+(${-k})\\times${-n}b$`;
-						texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${-k}(${printlatex(`a+(${-n})*b`)})$`;
-					}
-					break;
-				case 'nka+mkb':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${n * k}*a+(${m * k})*b`)}$`;
-					texte_corr = texte;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}\\times${n}a+${k}\\times${m}b$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}(${n}a+${m}b)$`;
-					break;
-				case 'nka-mkb':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${n * k}*a-(${m * k})*b`)}$`;
-					texte_corr = texte;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}\\times${n}a-${k}\\times${m}b$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}(${n}a-${m}b)$`;
-					break;
-				case 'nkx+mkx2':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${n * k}*x+(${m * k})*x^2`)}$`;
-					texte_corr = texte;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}x\\times${n}+${k}x\\times${m}x$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}x(${n}+${m}x)$`;
-					break;
-				case 'nkx-mkx2':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${printlatex(`${n * k}*x-(${m * k})*x^2`)}$`;
-					texte_corr = texte;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}x\\times${n}-${k}x\\times${m}x$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=${k}x(${n}-${m}x)$`;
-					break;
-				case 'nx2+x':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${n}x^2+x$`;
-					texte_corr = texte;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=x\\times ${n}x+x\\times 1$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=x(${n}x+1)$`;
-					break;
-				case 'nx2+mx':
-					texte = `$${lettre_depuis_chiffre(i + 1)}=${n}x^2+${m}x$`;
-					texte_corr = texte;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=x\\times ${n}x+x\\times ${m}$`;
-					texte_corr += `<br>$\\phantom{${lettre_depuis_chiffre(i + 1)}}=x(${n}x+${m})$`;
-					break;
-
-
-			}
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++;
-		}
-		liste_de_question_to_contenu(this);
-	}
-}
-
-/**
-* Déterminer l'image d'un nombre par une fonction d'après sa forme algébrique
-*
-* * Niveau 1 : Fonctions affines
-* * Niveau 2 : Polynôme du second degré
-* * Niveau 3 : Quotients de fonctions affines
-* * Niveau 4 : (ax+b)(cx+d)
-* * Niveau 5 : Mélange 
-* @Auteur Rémi Angot
-* 3F12-3
-*/
-function Tableau_de_valeurs() {
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Compléter un tableau de valeurs";
-	this.consigne = "";
-	this.nb_questions = 1;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.spacing = 1;
-	this.sup = 5; // niveau de difficulté
-	this.correction_detaillee_disponible = true;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-
-		let type_de_questions_disponibles = [];
-		if (this.sup == 1) {
-			type_de_questions_disponibles = ['ax+b', 'ax'];
-		}
-		if (this.sup == 2) {
-			type_de_questions_disponibles = ['ax2+bx+c', 'ax2+c', 'ax2+bx'];
-		}
-		if (this.sup == 3) {
-			type_de_questions_disponibles = ['a/cx+d', 'ax+b/cx+d'];
-		}
-		if (this.sup == 4) {
-			type_de_questions_disponibles = ['(ax+b)(cx+d)', '(ax+b)2'];
-		}
-		if (this.sup == 5) {
-			type_de_questions_disponibles = ['ax+b', 'ax', 'ax2+bx+c', 'ax2+c', 'ax2+bx', 'a/cx+d', 'ax+b/cx+d', '(ax+b)(cx+d)', '(ax+b)2']
-		}
-		let liste_type_de_questions = combinaison_listes(type_de_questions_disponibles, this.nb_questions); // Tous les types de questions sont posées mais l'ordre diffère à chaque "cycle"
-		let liste_de_x = combinaison_listes([[-3, 0, 3], [-2, 0, 2], [1, 2, 5], [-3, 6, 9]], this.nb_questions);
-		for (let i = 0, texte, texte_corr, a, b, c, d, x1, x2, x3, expression, nomdef, ligne2, calculs = "", cpt = 0; i < this.nb_questions && cpt < 50;) {
-			nomdef = lettre_minuscule_depuis_chiffre(6 + i) // on commence par f puis on continue dans l'ordre alphabétique
-			x1 = liste_de_x[i][0];
-			x2 = liste_de_x[i][1];
-			x3 = liste_de_x[i][2];
-			switch (liste_type_de_questions[i]) {
-				case 'ax+b':
-					a = randint(-10, 10, [0, -1, 1])
-					b = randint(-10, 10, [0])
-					expression = `${a}x${ecriture_algebrique(b)}`
-					ligne2 = `${nomdef}(x) & ${a * liste_de_x[i][0] + b} & ${a * liste_de_x[i][1] + b} & ${a * liste_de_x[i][2] + b} \\\\\n`
-					calculs = `$${nomdef}(${x1})=${a}\\times${ecriture_parenthese_si_negatif(x1)}${ecriture_algebrique(b)}=${a * x1}${ecriture_algebrique(b)}=${a * x1 + b}$<br>`
-					calculs += `$${nomdef}(${x2})=${a}\\times${ecriture_parenthese_si_negatif(x2)}${ecriture_algebrique(b)}=${a * x2}${ecriture_algebrique(b)}=${a * x2 + b}$<br>`
-					calculs += `$${nomdef}(${x3})=${a}\\times${ecriture_parenthese_si_negatif(x3)}${ecriture_algebrique(b)}=${a * x3}${ecriture_algebrique(b)}=${a * x3 + b}$<br>`
-					break;
-				case 'ax':
-					a = randint(-10, 10, [0, -1, 1])
-					expression = `${a}x`
-					ligne2 = `${nomdef}(x) & ${a * liste_de_x[i][0]} & ${a * liste_de_x[i][1]} & ${a * liste_de_x[i][2]} \\\\\n`
-					calculs = `$${nomdef}(${x1})=${a}\\times${ecriture_parenthese_si_negatif(x1)}=${a * x1}$<br>`
-					calculs += `$${nomdef}(${x2})=${a}\\times${ecriture_parenthese_si_negatif(x2)}=${a * x2}$<br>`
-					calculs += `$${nomdef}(${x3})=${a}\\times${ecriture_parenthese_si_negatif(x3)}=${a * x3}$<br>`
-					break;
-				case 'ax2+bx+c':
-					a = randint(-3, 3, [0, -1, 1])
-					b = randint(-5, 5, [0, -1, 1])
-					c = randint(-10, 10, [0])
-					expression = `${a}x^2${ecriture_algebrique(b)}x${ecriture_algebrique(c)}`
-					ligne2 = `${nomdef}(x) & ${a * liste_de_x[i][0] ** 2 + b * liste_de_x[i][0] + c} & ${a * liste_de_x[i][1] ** 2 + b * liste_de_x[i][1] + c} & ${a * liste_de_x[i][2] ** 2 + b * liste_de_x[i][2] + c} \\\\\n`
-					calculs = `$${nomdef}(${x1})=${a}\\times${ecriture_parenthese_si_negatif(x1)}^2${ecriture_algebrique(b)}\\times${ecriture_parenthese_si_negatif(x1)}${ecriture_algebrique(c)}=${a}\\times${x1 ** 2}${ecriture_algebrique(b * x1)}${ecriture_algebrique(c)}=${a * x1 ** 2 + b * x1 + c}$<br>`
-					calculs += `$${nomdef}(${x2})=${a}\\times${ecriture_parenthese_si_negatif(x2)}^2${ecriture_algebrique(b)}\\times${ecriture_parenthese_si_negatif(x2)}${ecriture_algebrique(c)}=${a}\\times${x2 ** 2}${ecriture_algebrique(b * x2)}${ecriture_algebrique(c)}=${a * x2 ** 2 + b * x2 + c}$<br>`
-					calculs += `$${nomdef}(${x3})=${a}\\times${ecriture_parenthese_si_negatif(x3)}^2${ecriture_algebrique(b)}\\times${ecriture_parenthese_si_negatif(x3)}${ecriture_algebrique(c)}=${a}\\times${x3 ** 2}${ecriture_algebrique(b * x3)}${ecriture_algebrique(c)}=${a * x3 ** 2 + b * x3 + c}$<br>`
-					break;
-				case 'ax2+c':
-					a = randint(-4, 4, [0, -1, 1])
-					c = randint(-10, 10, [0])
-					expression = `${a}x^2${ecriture_algebrique(c)}`
-					ligne2 = `${nomdef}(x) & ${a * liste_de_x[i][0] ** 2 + c} & ${a * liste_de_x[i][1] ** 2 + c} & ${a * liste_de_x[i][2] ** 2 + c} \\\\\n`
-					calculs = `$${nomdef}(${x1})=${a}\\times${ecriture_parenthese_si_negatif(x1)}^2${ecriture_algebrique(c)}=${a}\\times${x1 ** 2}${ecriture_algebrique(c)}=${a * x1 ** 2 + c}$<br>`
-					calculs += `$${nomdef}(${x2})=${a}\\times${ecriture_parenthese_si_negatif(x2)}^2${ecriture_algebrique(c)}=${a}\\times${x2 ** 2}${ecriture_algebrique(c)}=${a * x2 ** 2 + c}$<br>`
-					calculs += `$${nomdef}(${x3})=${a}\\times${ecriture_parenthese_si_negatif(x3)}^2${ecriture_algebrique(c)}=${a}\\times${x3 ** 2}${ecriture_algebrique(c)}=${a * x3 ** 2 + c}$<br>`
-					break;
-				case 'ax2+bx':
-					a = randint(-3, 3, [0, -1, 1])
-					b = randint(-5, 5, [0, -1, 1])
-					c = randint(-10, 10, [0])
-					expression = `${a}x^2${ecriture_algebrique(b)}x`
-					ligne2 = `${nomdef}(x) & ${a * liste_de_x[i][0] ** 2 + b * liste_de_x[i][0]} & ${a * liste_de_x[i][1] ** 2 + b * liste_de_x[i][1]} & ${a * liste_de_x[i][2] ** 2 + b * liste_de_x[i][2]} \\\\\n`
-					calculs = `$${nomdef}(${x1})=${a}\\times${ecriture_parenthese_si_negatif(x1)}^2${ecriture_algebrique(b)}\\times${ecriture_parenthese_si_negatif(x1)}=${a}\\times${x1 ** 2}${ecriture_algebrique(b * x1)}=${a * x1 ** 2 + b * x1}$<br>`
-					calculs += `$${nomdef}(${x2})=${a}\\times${ecriture_parenthese_si_negatif(x2)}^2${ecriture_algebrique(b)}\\times${ecriture_parenthese_si_negatif(x2)}=${a}\\times${x2 ** 2}${ecriture_algebrique(b * x2)}=${a * x2 ** 2 + b * x2}$<br>`
-					calculs += `$${nomdef}(${x3})=${a}\\times${ecriture_parenthese_si_negatif(x3)}^2${ecriture_algebrique(b)}\\times${ecriture_parenthese_si_negatif(x3)}=${a}\\times${x3 ** 2}${ecriture_algebrique(b * x3)}=${a * x3 ** 2 + b * x3}$<br>`
-					break;
-				case 'a/cx+d':
-					a = randint(-10, 10, [0])
-					c = randint(-10, 10, [0, -1, 1])
-					d = randint(-10, 10, [0])
-					while (c * x1 + d == 0 || c * x2 + d == 0 || c * x3 + d == 0) {
-						c = randint(-10, 10, [0, -1, 1])
-					}
-					expression = `\\dfrac{${a}}{${c}x${ecriture_algebrique(d)}}`
-					ligne2 = `${nomdef}(x) & ${tex_fraction_reduite(a, c * liste_de_x[i][0] + d)} & ${tex_fraction_reduite(a, c * liste_de_x[i][1] + d)} & ${tex_fraction_reduite(a, c * liste_de_x[i][2] + d)} \\\\\n`
-					calculs = `$${nomdef}(${x1})=\\dfrac{${a}}{${c}\\times${ecriture_parenthese_si_negatif(x1)}${ecriture_algebrique(d)}}=\\dfrac{${a}}{${c * x1}${ecriture_algebrique(d)}}=\\dfrac{${a}}{${c * x1 + d}}`
-					if (pgcd(a, c * x1 + d) == 1) {
-						calculs += '$<br><br>'
-					} else {
-						calculs += '=' + tex_fraction_reduite(a, c * x1 + d) + '$<br><br>'
-					}
-					calculs += `$${nomdef}(${x2})=\\dfrac{${a}}{${c}\\times${ecriture_parenthese_si_negatif(x2)}${ecriture_algebrique(d)}}=\\dfrac{${a}}{${c * x2}${ecriture_algebrique(d)}}=\\dfrac{${a}}{${c * x2 + d}}`
-					if (pgcd(a, c * x2 + d) == 1) {
-						calculs += '$<br><br>'
-					} else {
-						calculs += '=' + tex_fraction_reduite(a, c * x2 + d) + '$<br><br>'
-					}
-					calculs += `$${nomdef}(${x3})=\\dfrac{${a}}{${c}\\times${ecriture_parenthese_si_negatif(x3)}${ecriture_algebrique(d)}}=\\dfrac{${a}}{${c * x3}${ecriture_algebrique(d)}}=\\dfrac{${a}}{${c * x3 + d}}`
-					if (pgcd(a, c * x3 + d) == 1) {
-						calculs += '$<br><br>'
-					} else {
-						calculs += '=' + tex_fraction_reduite(a, c * x3 + d) + '$<br><br>'
-					}
-					break;
-				case 'ax+b/cx+d':
-					a = randint(-10, 10, [0, 1, -1])
-					b = randint(-10, 10, [0])
-					c = randint(-10, 10, [0, -1, 1])
-					d = randint(-10, 10, [0])
-					while (c * x1 + d == 0 || c * x2 + d == 0 || c * x3 + d == 0) {
-						c = randint(-10, 10, [0, -1, 1])
-					}
-					expression = `\\dfrac{${a}x${ecriture_algebrique(b)}}{${c}x${ecriture_algebrique(d)}}`
-					ligne2 = `${nomdef}(x) & ${tex_fraction_reduite(a * liste_de_x[i][0] + b, c * liste_de_x[i][0] + d)} & ${tex_fraction_reduite(a * liste_de_x[i][1] + b, c * liste_de_x[i][1] + d)} & ${tex_fraction_reduite(a * liste_de_x[i][2] + b, c * liste_de_x[i][2] + d)} \\\\\n`
-					calculs = `$${nomdef}(${x1})=\\dfrac{${a}\\times${ecriture_parenthese_si_negatif(x1)}${ecriture_algebrique(b)}}{${c}\\times${ecriture_parenthese_si_negatif(x1)}${ecriture_algebrique(d)}}=\\dfrac{${a * x1}${ecriture_algebrique(b)}}{${c * x1}${ecriture_algebrique(d)}}=\\dfrac{${a * x1 + b}}{${c * x1 + d}}`
-					if (pgcd(a * x1 + b, c * x1 + d) == 1) {
-						calculs += '$<br><br>'
-					} else {
-						calculs += '=' + tex_fraction_reduite(a * x1 + b, c * x1 + d) + '$<br><br>'
-					}
-					calculs += `$${nomdef}(${x2})=\\dfrac{${a}\\times${ecriture_parenthese_si_negatif(x2)}${ecriture_algebrique(b)}}{${c}\\times${ecriture_parenthese_si_negatif(x2)}${ecriture_algebrique(d)}}=\\dfrac{${a * x2}${ecriture_algebrique(b)}}{${c * x2}${ecriture_algebrique(d)}}=\\dfrac{${a * x2 + b}}{${c * x2 + d}}`
-					if (pgcd(a * x2 + b, c * x2 + d) == 1) {
-						calculs += '$<br><br>'
-					} else {
-						calculs += '=' + tex_fraction_reduite(a * x2 + b, c * x2 + d) + '$<br><br>'
-					}
-					calculs += `$${nomdef}(${x3})=\\dfrac{${a}\\times${ecriture_parenthese_si_negatif(x3)}${ecriture_algebrique(b)}}{${c}\\times${ecriture_parenthese_si_negatif(x3)}${ecriture_algebrique(d)}}=\\dfrac{${a * x3}${ecriture_algebrique(b)}}{${c * x3}${ecriture_algebrique(d)}}=\\dfrac{${a * x3 + b}}{${c * x3 + d}}`
-					if (pgcd(a * x3 + b, c * x3 + d) == 1) {
-						calculs += '$<br><br>'
-					} else {
-						calculs += '=' + tex_fraction_reduite(a * x3 + b, c * x3 + d) + '$<br><br>'
-					}
-					break;
-				case '(ax+b)(cx+d)':
-					a = randint(-5, 5, [0, 1, -1])
-					b = randint(-5, 5, [0])
-					c = randint(-3, 3, [0, -1, 1])
-					d = randint(-3, 3, [0])
-					if (a < 0 && b < 0 && c < 0 && d < 0) {
-						d = randint(1, 3)
-					}
-					expression = `(${a}x${ecriture_algebrique(b)})(${c}x${ecriture_algebrique(d)})`
-					ligne2 = `${nomdef}(x) & ${(a * liste_de_x[i][0] + b) * (c * liste_de_x[i][0] + d)} & ${(a * liste_de_x[i][1] + b) * (c * liste_de_x[i][1] + d)} & ${(a * liste_de_x[i][2] + b) * (c * liste_de_x[i][2] + d)} \\\\\n`
-					calculs = `$${nomdef}(${x1})=\\left(${a}\\times${ecriture_parenthese_si_negatif(x1)}${ecriture_algebrique(b)}\\right)\\left(${c}\\times${ecriture_parenthese_si_negatif(x1)}${ecriture_algebrique(d)}\\right)=(${a * x1}${ecriture_algebrique(b)})(${c * x1}${ecriture_algebrique(d)})=${a * x1 + b}\\times ${ecriture_parenthese_si_negatif(c * x1 + d)}=${(a * x1 + b) * (c * x1 + d)}$<br>`
-					calculs += `$${nomdef}(${x2})=\\left(${a}\\times${ecriture_parenthese_si_negatif(x2)}${ecriture_algebrique(b)}\\right)\\left(${c}\\times${ecriture_parenthese_si_negatif(x2)}${ecriture_algebrique(d)}\\right)=(${a * x2}${ecriture_algebrique(b)})(${c * x2}${ecriture_algebrique(d)})=${a * x2 + b}\\times ${ecriture_parenthese_si_negatif(c * x2 + d)}=${(a * x2 + b) * (c * x2 + d)}$<br>`
-					calculs += `$${nomdef}(${x3})=\\left(${a}\\times${ecriture_parenthese_si_negatif(x3)}${ecriture_algebrique(b)}\\right)\\left(${c}\\times${ecriture_parenthese_si_negatif(x3)}${ecriture_algebrique(d)}\\right)=(${a * x3}${ecriture_algebrique(b)})(${c * x3}${ecriture_algebrique(d)})=${a * x3 + b}\\times ${ecriture_parenthese_si_negatif(c * x3 + d)}=${(a * x3 + b) * (c * x3 + d)}$<br>`
-					break;
-				case '(ax+b)2':
-					a = randint(-3, 3, [0, 1, -1])
-					b = randint(-3, 3, [0])
-					expression = `(${a}x${ecriture_algebrique(b)})^2`
-					ligne2 = `${nomdef}(x) & ${(a * liste_de_x[i][0] + b) ** 2} & ${(a * liste_de_x[i][1] + b) ** 2} & ${(a * liste_de_x[i][2] + b) ** 2} \\\\\n`
-					calculs = `$${nomdef}(${x1})=\\left(${a}\\times${ecriture_parenthese_si_negatif(x1)}${ecriture_algebrique(b)}\\right)^2=(${a * x1}${ecriture_algebrique(b)})^2=${ecriture_parenthese_si_negatif(a * x1 + b)}^2=${(a * x1 + b) ** 2}$<br>`
-					calculs += `$${nomdef}(${x2})=\\left(${a}\\times${ecriture_parenthese_si_negatif(x2)}${ecriture_algebrique(b)}\\right)^2=(${a * x2}${ecriture_algebrique(b)})^2=${ecriture_parenthese_si_negatif(a * x2 + b)}^2=${(a * x2 + b) ** 2}$<br>`
-					calculs += `$${nomdef}(${x3})=\\left(${a}\\times${ecriture_parenthese_si_negatif(x3)}${ecriture_algebrique(b)}\\right)^2=(${a * x3}${ecriture_algebrique(b)})^2=${ecriture_parenthese_si_negatif(a * x3 + b)}^2=${(a * x3 + b) ** 2}$<br>`
-
-					break;
-			}
-
-
-			texte = `On considère la fonction $${nomdef}$ définie par $${nomdef}:x\\mapsto ${expression}$. Compléter le tableau de valeurs suivant.`
-			texte_corr = ''
-			texte += `<br><br>`
-			if (sortie_html) {
-				texte += `$\\def\\arraystretch{2.5}\\begin{array}{|l|c|c|c|}\n`
-			} else {
-				texte += `$\\begin{array}{|l|c|c|c|}\n`
-			}
-
-			texte += `\\hline\n`
-			texte += `x & ${liste_de_x[i][0]} & ${liste_de_x[i][1]} & ${liste_de_x[i][2]} \\\\\n`
-			texte += `\\hline\n`
-			texte += `${nomdef}(x) & \\phantom{-10} & \\phantom{-10} & \\phantom{-10} \\\\\n`
-			texte += `\\hline\n`
-			texte += `\\end{array}\n$`
-
-
-			if (sortie_html) {
-				texte_corr = `$\\def\\arraystretch{2.5}\\begin{array}{|l|c|c|c|}\n`
-			} else {
-				texte_corr = `$\\begin{array}{|l|c|c|c|}\n`
-			}
-
-			texte_corr += `\\hline\n`;
-			texte_corr += `x & ${liste_de_x[i][0]} & ${liste_de_x[i][1]} & ${liste_de_x[i][2]} \\\\\n`;
-			texte_corr += `\\hline\n`;
-			texte_corr += ligne2;
-			texte_corr += `\\hline\n`;
-			texte_corr += `\\end{array}\n$`;
-			if (this.correction_detaillee) {
-				texte_corr += '<br><br>';
-				texte_corr += calculs;
-			}
-
-
-
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++;
-		}
-		this.nb_questions == 1 ? liste_de_question_to_contenu_sans_numero(this) : liste_de_question_to_contenu(this);
-	}
-	this.besoin_formulaire_numerique = ['Niveau de difficulté', 5, '1 : Fonctions affines\n2 : Polynome du second degré\n3 : Quotient\n4 : Produit \n5 : Mélange'];
-}
-
-/**
- * Développer des expressions de la forme(ax+ou-b)(cx+ou-d)
-* @auteur Jean-Claude Lhote
-* 3L11-1
-*/
-function Double_distributivite() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Utiliser la double distributivité";
-	this.consigne = "Développer et réduire les expressions suivantes.";
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.spacing = 1;
-	this.spacing_corr = 1;
-	this.nb_questions = 5;
-	this.sup = 1;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		let type_de_questions_disponibles = [1, 2];
-		if (this.sup == 2) {
-			type_de_questions_disponibles = [3, 4]
-		}
-		if (this.sup == 3) {
-			type_de_questions_disponibles = [1, 2, 3, 4]
-		}
-
-
-		let liste_type_de_questions = combinaison_listes(type_de_questions_disponibles, this.nb_questions)
-		for (let i = 0, texte, texte_corr, cpt = 0, a, b, c, d, type_de_questions; i < this.nb_questions && cpt < 50;) {
-			type_de_questions = liste_type_de_questions[i];
-			a = randint(2, 9);
-			b = randint(2, 9);
-			c = randint(2, 9, [a]);
-			d = randint(2, 9, [b]);
-			switch (type_de_questions) {
-				case 1: //(x+b)(x+d)
-					b = randint(2, 10)
-					d = randint(2, 12)
-					texte = `$(x+${b})(x+${d})$`
-					texte_corr = `$(x+${b})(x+${d})=x^2+${b}x+${d}x+${b * d}=x^2+${b + d}x+${b * d}$`
-					break;
-				case 2: //(ax+b)(cx+d)
-					texte = `$(${a}x+${b})(${c}x+${d})$`
-					texte_corr = `$(${a}x+${b})(${c}x+${d})=${a * c}x^2+${a * d}x+${b * c}x+${b * d}=${a * c}x^2+${a * d + b * c}x+${b * d}$`
-					break;
-				case 3://(ax-b)(cx+d)
-					texte = `$(${a}x-${b})(${c}x+${d})$`
-					if (egal(a * d - b * c, 0))
-						texte_corr = `$(${a}x-${b})(${c}x+${d})=${a * c}x^2+${d * a}x-${b * c}x-${b * d}=${printlatex(`${a * c}*x^2-${b * d}`)}$`;
-					else texte_corr = `$(${a}x-${b})(${c}x+${d})=${a * c}x^2+${d * a}x-${b * c}x-${b * d}=${printlatex(`${a * c}*x^2+(${d * a - b * c})*x-${b * d}`)}$`;
-					break;
-				case 4://(ax-b)(cx-d)
-					texte = `$(${a}x-${b})(${c}x-${d})$`
-					texte_corr = `$(${a}x-${b})(${c}x-${d})=${a * c}x^2-${a * d}x-${b * c}x+${b * d}=${a * c}x^2-${a * d + b * c}x+${b * d}$`
-					break;
-			}
-			if (this.liste_questions.indexOf(texte) == -1) {
-				// Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++;
-		}
-		liste_de_question_to_contenu(this);
-	}
-	this.besoin_formulaire_numerique = ['Niveau de difficulté', 3, '1 : (x+a)(x+b) et (ax+b)(cx+d)\n 2 : (ax-b)(cx+d) et (ax-b)(cx-d)\n 3 : Tous les types'];
-}
-
-/**
- * Développer (ax-b)(ax+b)
-* @auteur Jean-Claude Lhote
-* 3L12-1
-*/
-function Developper_Identites_remarquables3() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Développer (a-b)(a+b)";
-	this.consigne = "Développer les expressions suivantes.";
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.spacing = 1;
-	this.spacing_corr = 1;
-	this.nb_questions = 5;
-	this.sup = 2;
-
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		let liste_fractions = [[1, 2], [1, 3], [2, 3], [1, 4], [3, 4], [1, 5], [2, 5], [3, 5], [4, 5],
-		[1, 6], [5, 6], [1, 7], [2, 7], [3, 7], [4, 7], [5, 7], [6, 7], [1, 8], [3, 8], [5, 8], [7, 8],
-		[1, 9], [2, 9], [4, 9], [5, 9], [7, 9], [8, 9], [1, 10], [3, 10], [7, 10], [9, 10]]
-		for (let i = 0, ns, ds, texte, texte_corr, cpt = 0, a, b, fraction = []; i < this.nb_questions && cpt < 50;) {
-			if (this.sup == 1) {
-				a = randint(1, 9);	 // coef de x est égal à 1
-				texte = `$(x-${a})(x+${a})$`    // (x-a)(x+a)
-				texte_corr = `$(x-${a})(x+${a})=x^2-${a}^2=x^2-${a * a}$`;
-			}
-			else if (this.sup == 2) {
-				a = randint(1, 9)  // (bx-a)(bx+a) avec a et b entier positifs entre 1 et 9,  b différent de 1
-				b = randint(2, 9);
-				texte = `$(${b}x-${a})(${b}x+${a})$`; // b>1
-				texte_corr = `$(${b}x-${a})(${b}x+${a})=(${b}x)^2-${a}^2=${b * b}x^2-${a * a}$`;
-			}
-			else {   //  (bx-a)(bx+a) avec a entier et b rationnel simple
-				a = randint(1, 9);
-				fraction = choice(liste_fractions);
-				ns = fraction[0]
-				ds = fraction[1]
-				texte = `$\\left(${tex_fraction(ns, ds)}x-${a}\\right)\\left(${tex_fraction(ns, ds)}x+${a}\\right)$`; // b>1
-				texte_corr = `$\\left(${tex_fraction(ns, ds)}x-${a}\\right)\\left(${tex_fraction(ns, ds)}x+${a}\\right)=\\left(${tex_fraction(ns, ds)}x\\right)^2-${a}^2=${tex_fraction(ns * ns, ds * ds)}x^2-${a * a}$`;
-			}
-
-			if (this.liste_questions.indexOf(texte) == -1) {
-				// Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++;
-		}
-		liste_de_question_to_contenu(this);
-	}
-	this.besoin_formulaire_numerique = ['Niveau de difficulté', 3, '1 : Coefficient de x égal à 1\n 2 : Coefficient de x supérieur à 1\n 3 : Coefficient de x rationnel'];
-}
-
-/**
- * Factoriser a²-b²
-* @auteur Jean-Claude Lhote
-* 3L12
-*/
-function Factoriser_Identites_remarquables3() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Factoriser a²-b²";
-	this.consigne = "Factoriser les expressions suivantes.";
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.spacing = 1;
-	this.spacing_corr = 1;
-	this.nb_questions = 5;
-	this.sup = 2;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		let liste_fractions = [[1, 2], [1, 3], [2, 3], [1, 4], [3, 4], [1, 5], [2, 5], [3, 5], [4, 5],
-		[1, 6], [5, 6], [1, 7], [2, 7], [3, 7], [4, 7], [5, 7], [6, 7], [1, 8], [3, 8], [5, 8], [7, 8],
-		[1, 9], [2, 9], [4, 9], [5, 9], [7, 9], [8, 9], [1, 10], [3, 10], [7, 10], [9, 10]]
-		for (let i = 0, texte, texte_corr, cpt = 0, a, b, ns, ds, fraction = []; i < this.nb_questions && cpt < 50;) {
-			if (this.sup == 1) {
-				a = randint(1, 9);	 // coef de x est égal à 1
-				texte = `$x^2-${a * a}$`    // (x-a)(x+a)
-				texte_corr = `$x^2-${a * a}=x^2-${a}^2=(x-${a})(x+${a})$`;
-			}
-			else if (this.sup == 2) {
-				a = randint(1, 9)  // (bx-a)(bx+a) avec a et b entier positifs entre 1 et 9,  b différent de 1
-				b = randint(2, 9);
-				texte = `$${b * b}x^2-${a * a}$`; // b>1
-				texte_corr = `$${b * b}x^2-${a * a}=(${b}x)^2-${a}^2=(${b}x-${a})(${b}x+${a})$`;
-			}
-			else {   //  (bx-a)(bx+a) avec a entier et b rationnel simple
-				a = randint(1, 9);
-				fraction = choice(liste_fractions);
-				ns = fraction[0]
-				ds = fraction[1]
-				texte = `$${tex_fraction(ns * ns, ds * ds)}x^2-${a * a}$`; // b>1
-				texte_corr = `$${tex_fraction(ns * ns, ds * ds)}x^2-${a * a}=\\left(${tex_fraction(ns, ds)}x\\right)^2-${a}^2=\\left(${tex_fraction(ns, ds)}x-${a}\\right)\\left(${tex_fraction(ns, ds)}x+${a}\\right)$`;
-
-			}
-
-			if (this.liste_questions.indexOf(texte) == -1) {
-				// Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++;
-		}
-		liste_de_question_to_contenu(this);
-	}
-	this.besoin_formulaire_numerique = ['Niveau de difficulté', 3, '1 : Coefficient de x égal à 1\n 2 : Coefficient de x supérieur à 1\n 3 : Coefficient de x rationnel'];
-}
-
-
-
-/**
-* Équation du premier degré
-* * Type 1 : ax+b=cx+d
-* * Type 2 : k(ax+b)=cx+d
-* * Type 3 : k-(ax+b)=cx+d
-* * Tous les types
-* @Auteur Rémi Angot
-* 3L13-1
-*/
-function Exercice_equation1_2() {
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Équation du premier degré (utilisant la distributivité)"
-	this.consigne = 'Résoudre les équations suivantes'
-	this.spacing = 2;
-	sortie_html ? this.spacing_corr = 3 : this.spacing_corr = 2;
-	this.correction_detaillee_disponible = true;
-	if (!sortie_html) {
-		this.correction_detaillee = false;
-	}
-	this.nb_questions = 3;
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		let liste_type_de_questions = ['ax+b=cx+d', 'k(ax+b)=cx+d', 'k-(ax+b)=cx+d'];
-		liste_type_de_questions = combinaison_listes(liste_type_de_questions, this.nb_questions)
-		for (let i = 0, a, b, c, d, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) { // On limite le nombre d'essais pour chercher des valeurs nouvelles
-			a = randint(-9, 9, 0);
-			b = randint(-9, 9, 0);
-			c = randint(-9, 9, 0);
-			d = randint(-9, 9, 0);
-			k = randint(2, 9)
-			if (liste_type_de_questions[i] == 'ax+b=cx+d') {
-				if (c == a) { c = randint(1, 9, [a]) } // sinon on arrive à une division par 0
-				if (!this.sup && a < c) {
-					c = randint(1, 9)
-					a = randint(c + 1, 15) // a sera plus grand que c pour que a-c>0
-				}
-				texte = `$${rien_si_1(a)}x${ecriture_algebrique(b)}=${rien_si_1(c)}x${ecriture_algebrique(d)}$`;
-				texte_corr = texte + '<br>';
-				if (this.correction_detaillee) {
-					if (c > 0) {
-						texte_corr += `On soustrait $${rien_si_1(c)}x$ aux deux membres.<br>`
-					} else {
-						texte_corr += `On ajoute $${rien_si_1(-1 * c)}x$ aux deux membres.<br>`
-					}
-				}
-				texte_corr += `$${rien_si_1(a)}x${ecriture_algebrique(b)}${mise_en_evidence(signe(-1 * c) + rien_si_1(abs(c)) + 'x')}=${c}x+${d}${mise_en_evidence(signe(-1 * c) + rien_si_1(abs(c)) + 'x')}$<br>`;
-				texte_corr += `$${rien_si_1(a - c)}x${ecriture_algebrique(b)}=${d}$<br>`
-				if (this.correction_detaillee) {
-					if (b > 0) {
-						texte_corr += `On soustrait $${b}$ aux deux membres.<br>`
-					} else {
-						texte_corr += `On ajoute $${-1 * b}$ aux deux membres.<br>`
-					}
-				}
-				texte_corr += `$${rien_si_1(a - c)}x${ecriture_algebrique(b)}${mise_en_evidence(ecriture_algebrique(-1 * b))}=${d}${mise_en_evidence(ecriture_algebrique(-1 * b))}$<br>`
-				texte_corr += `$${rien_si_1(a - c)}x=${d - b}$<br>`
-
-				if (this.correction_detaillee) { texte_corr += `On divise les deux membres par $${a - c}$.<br>` }
-				texte_corr += `$${rien_si_1(a - c)}x${mise_en_evidence('\\div' + ecriture_parenthese_si_negatif(a - c))}=${d - b + mise_en_evidence('\\div' + ecriture_parenthese_si_negatif(a - c))}$<br>`
-				texte_corr += `$x=${tex_fraction(d - b, a - c)}$`
-				if (pgcd(abs(d - b), abs(a - c)) > 1 || (a - c) < 0) {
-					texte_corr += `<br>$x=${tex_fraction_reduite(d - b, a - c)}$`
-				}
-				texte_corr += `<br> La solution est $${tex_fraction_reduite(d - b, a - c)}$.`
-			}
-
-			if (liste_type_de_questions[i] == 'k(ax+b)=cx+d') {
-				if (c == k * a) { c = randint(1, 9, [a]) } // sinon on arrive à une division par 0
-				texte = `$${k}(${rien_si_1(a)}x${ecriture_algebrique(b)})=${rien_si_1(c)}x${ecriture_algebrique(d)}$`;
-				texte_corr = texte + '<br>';
-				if (this.correction_detaillee) {
-					texte_corr += 'On développe le membre de gauche.<br>'
-				}
-				texte_corr += `$${k * a}x${ecriture_algebrique(k * b)}=${rien_si_1(c)}x${ecriture_algebrique(d)}$<br>`;
-				if (this.correction_detaillee) {
-					if (c > 0) {
-						texte_corr += `On soustrait $${rien_si_1(c)}x$ aux deux membres.<br>`
-					} else {
-						texte_corr += `On ajoute $${rien_si_1(-1 * c)}x$ aux deux membres.<br>`
-					}
-				}
-				texte_corr += `$${k * a}x${ecriture_algebrique(k * b)}${mise_en_evidence(signe(-1 * c) + rien_si_1(abs(c)) + 'x')}=${c}x${ecriture_algebrique(d)}${mise_en_evidence(signe(-1 * c) + rien_si_1(abs(c)) + 'x')}$<br>`;
-				texte_corr += `$${rien_si_1(k * a - c)}x${ecriture_algebrique(k * b)}=${d}$<br>`
-				if (this.correction_detaillee) {
-					if (k * b > 0) {
-						texte_corr += `On soustrait $${k * b}$ aux deux membres.<br>`
-					} else {
-						texte_corr += `On ajoute $${-k * b}$ aux deux membres.<br>`
-					}
-				}
-				texte_corr += `$${rien_si_1(k * a - c)}x${ecriture_algebrique(k * b)}${mise_en_evidence(ecriture_algebrique(-k * b))}=${d}${mise_en_evidence(ecriture_algebrique(-k * b))}$<br>`
-				texte_corr += `$${rien_si_1(k * a - c)}x=${d - k * b}$<br>`
-
-				if (this.correction_detaillee) { texte_corr += `On divise les deux membres par $${k * a - c}$.<br>` }
-				texte_corr += `$${rien_si_1(k * a - c)}x${mise_en_evidence('\\div' + ecriture_parenthese_si_negatif(k * a - c))}=${d - k * b + mise_en_evidence('\\div' + ecriture_parenthese_si_negatif(k * a - c))}$<br>`
-				texte_corr += `$x=${tex_fraction(d - k * b, k * a - c)}$`
-				if (pgcd(abs(d - k * b), abs(k * a - c)) > 1 || (k * a - c) < 0) {
-					texte_corr += `<br>$x=${tex_fraction_reduite(d - k * b, k * a - c)}$`
-				}
-				texte_corr += `<br> La solution est $${tex_fraction_reduite(d - k * b, k * a - c)}$.`
-			}
-
-			if (liste_type_de_questions[i] == 'k-(ax+b)=cx+d') {
-				if (c == -a) { c = randint(-9, 9, [0, a]) } // sinon on arrive à une division par 0
-				texte = `$${k}-(${rien_si_1(a)}x${ecriture_algebrique(b)})=${rien_si_1(c)}x${ecriture_algebrique(d)}$`;
-				texte_corr = texte + '<br>';
-				if (this.correction_detaillee) {
-					texte_corr += 'On développe le membre de gauche.<br>'
-				}
-				texte_corr += `$${k}${ecriture_algebrique(-a)}x${ecriture_algebrique(-b)}=${rien_si_1(c)}x${ecriture_algebrique(d)}$<br>`;
-				texte_corr += `$${rien_si_1(-a)}x${ecriture_algebrique(k - b)}=${rien_si_1(c)}x${ecriture_algebrique(d)}$<br>`;
-
-				//On reprend le cas ax+b=cx+d en changeant les valeurs de a et b
-				a = -a;
-				b = k - b;
-
-				if (this.correction_detaillee) {
-					if (c > 0) {
-						texte_corr += `On soustrait $${rien_si_1(c)}x$ aux deux membres.<br>`
-					} else {
-						texte_corr += `On ajoute $${rien_si_1(-1 * c)}x$ aux deux membres.<br>`
-					}
-				}
-				texte_corr += `$${rien_si_1(a)}x${ecriture_algebrique(b)}${mise_en_evidence(signe(-1 * c) + rien_si_1(abs(c)) + 'x')}=${c}x+${d}${mise_en_evidence(signe(-1 * c) + rien_si_1(abs(c)) + 'x')}$<br>`;
-				texte_corr += `$${rien_si_1(a - c)}x${ecriture_algebrique(b)}=${d}$<br>`
-				if (this.correction_detaillee) {
-					if (b > 0) {
-						texte_corr += `On soustrait $${b}$ aux deux membres.<br>`
-					} else {
-						texte_corr += `On ajoute $${-1 * b}$ aux deux membres.<br>`
-					}
-				}
-				texte_corr += `$${rien_si_1(a - c)}x${ecriture_algebrique(b)}${mise_en_evidence(ecriture_algebrique(-1 * b))}=${d}${mise_en_evidence(ecriture_algebrique(-1 * b))}$<br>`
-				texte_corr += `$${rien_si_1(a - c)}x=${d - b}$<br>`
-
-				if (this.correction_detaillee) { texte_corr += `On divise les deux membres par $${a - c}$.<br>` }
-				texte_corr += `$${rien_si_1(a - c)}x${mise_en_evidence('\\div' + ecriture_parenthese_si_negatif(a - c))}=${d - b + mise_en_evidence('\\div' + ecriture_parenthese_si_negatif(a - c))}$<br>`
-				texte_corr += `$x=${tex_fraction(d - b, a - c)}$`
-				if (pgcd(abs(d - b), abs(a - c)) > 1 || (a - c) < 0) {
-					texte_corr += `<br>$x=${tex_fraction_reduite(d - b, a - c)}$`
-				}
-				texte_corr += `<br> La solution est $${tex_fraction_reduite(d - b, a - c)}$.`
-			}
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte)//replace(/1x/g,'x')); //remplace 1x par x
-				this.liste_corrections.push(texte_corr) //.replace(/1x/g,'x')); //remplace 1x par x
-				i++;
-			}
-			cpt++;
-		}
-		liste_de_question_to_contenu(this);
-	}
-}
-
-
-/**
- * Résolution d'équations de type (ax+b)(cx+d)=0
-* @auteur Jean-Claude Lhote
-* Tout est dans le nom de la fonction.
-* 3L14
-*/
-function Resoudre_une_equation_produit_nul() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Résoudre une équation produit nul";
-	this.consigne = "Résoudre les équations suivantes";
-	this.nb_questions = 5;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.sup = 1;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1.5
-	this.spacing = 1
-
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		let liste_fractions = [[1, 2], [1, 3], [2, 3], [1, 4], [3, 4], [1, 5], [2, 5], [3, 5], [4, 5],
-		[1, 6], [5, 6], [1, 7], [2, 7], [3, 7], [4, 7], [5, 7], [6, 7], [1, 8], [3, 8], [5, 8], [7, 8],
-		[1, 9], [2, 9], [4, 9], [5, 9], [7, 9], [8, 9], [1, 10], [3, 10], [7, 10], [9, 10]]
-		let liste_type_de_questions = []
-		switch (parseInt(this.sup)) {
-			case 1: liste_type_de_questions = combinaison_listes([1, 2], this.nb_questions);
-				break;
-			case 2: liste_type_de_questions = combinaison_listes([3, 4], this.nb_questions);
-				break;
-			case 3: liste_type_de_questions = combinaison_listes([5, 6], this.nb_questions);
-				break;
-			case 4: liste_type_de_questions = combinaison_listes([1, 2, 3, 4, 5, 6], this.nb_questions);
-
-		}
-		for (let i = 0, a, b, c, d, fraction1, fraction2, ns1, ns2, ds1, ds2, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			fraction1 = choice(liste_fractions);
-			ns1 = fraction1[0]
-			ds1 = fraction1[1]
-			fraction2 = choice(liste_fractions);
-			ns2 = fraction2[0]
-			ds2 = fraction2[1]
-			switch (liste_type_de_questions[i]) {
-				case 1: b = randint(1, 20); // (x+a)(x+b)=0 avec a et b entiers
-					d = randint(1, 20, [b])
-					texte = `$(x+${b})(x+${d})=0$`
-					texte_corr = 'Un produit est nul si l\'un au moins de ses facteurs est nul.'
-					texte_corr += '<br>' + `$(x+${b})(x+${d})=0$`
-					texte_corr += '<br> Soit ' + `$x+${b}=0$` + ' ou ' + `$x+${d}=0$`
-					texte_corr += '<br> Donc ' + `$x=${0 - b}$` + ' ou ' + `$x=${0 - d}$`
-					break;
-				case 2: b = randint(1, 20); // (x-a)(x+b)=0 avec a et b entiers
-					d = randint(1, 20, [b])
-					texte = `$(x-${b})(x+${d})=0$`
-					texte_corr = 'Un produit est nul si l\'un au moins de ses facteurs est nul.'
-					texte_corr += '<br>' + `$(x-${b})(x+${d})=0$`
-					texte_corr += '<br> Soit ' + `$x-${b}=0$` + ' ou  ' + `$x+${d}=0$`
-					texte_corr += '<br> Donc ' + `$x=${b}$` + ' ou ' + `$x=${0 - d}$`
-					break;
-
-				case 3: a = randint(2, 6); 	//(ax+b)(cx+d)=0  avec b/a et d/c entiers.
-					b = Math.round(randint(1, 5) * a);
-					c = randint(2, 6, [a]);
-					d = Math.round(randint(1, 5) * c);
-					texte = `$(${a}x+${b})(${c}x+${d})=0$`
-					texte_corr = 'Un produit est nul si l\'un au moins de ses facteurs est nul.'
-					texte_corr += '<br>' + `$(${a}x+${b})(${c}x+${d})=0$`
-					texte_corr += '<br> Soit ' + `$${a}x+${b}=0$` + ' ou ' + `$${c}x+${d}=0$`
-					texte_corr += '<br> Donc ' + `$${a}x=${0 - b}$` + ' ou ' + `$${c}x=${0 - d}$`
-					texte_corr += '<br> Donc ' + `$x=-${tex_fraction(b, a)}$` + ' ou ' + `$x=-${tex_fraction(d, c)}$`
-					texte_corr += '<br> Donc ' + `$x=${0 - b / a}$` + ' ou ' + `$x=${0 - d / c}$`
-					break;
-				case 4: a = randint(2, 6); 	//(ax+b)(cx-d)=0  avec b/a et d/c entiers.
-					b = Math.round(randint(1, 5) * a);
-					c = randint(2, 6, [a]);
-					d = Math.round(randint(1, 5) * c);
-					texte = `$(${a}x+${b})(${c}x-${d})=0$`
-					texte_corr = 'Un produit est nul si l\'un au moins de ses facteurs est nul.'
-					texte_corr += '<br>' + `$(${a}x+${b})(${c}x-${d})=0$`
-					texte_corr += '<br> Soit ' + `$${a}x+${b}=0$` + ' ou ' + `$${c}x-${d}=0$`
-					texte_corr += '<br> Donc ' + `$${a}x=${0 - b}$` + ' ou ' + `$${c}x=${d}$`
-					texte_corr += '<br> Donc ' + `$x=-${tex_fraction(b, a)}$` + ' ou ' + `$x=${tex_fraction(d, c)}$`
-					texte_corr += '<br> Donc ' + `$x=${0 - b / a}$` + ' ou ' + `$x=${d / c}$`
-					break;
-				case 5:
-					a = randint(2, 9);	//(ax+b)(cx+d)=0 	avec b/a et d/c quelconques.
-					b = randint(1, 20, [a]);
-					c = randint(2, 9, [a]);
-					d = randint(1, 20, [b, c]);
-					texte = `$(${a}x+${b})(${c}x+${d})=0$`
-					texte_corr = 'Un produit est nul si l\'un au moins de ses facteurs est nul.'
-					texte_corr += '<br>' + `$(${a}x+${b})(${c}x+${d})=0$`
-					texte_corr += '<br> Soit ' + `$${a}x+${b}=0$` + ' ou ' + `$${c}x+${d}=0$`
-					texte_corr += '<br> Donc ' + `$${a}x=${0 - b}$` + ' ou ' + `$${c}x=${0 - d}$`
-					texte_corr += '<br> Donc ' + `$x=-${tex_fraction(b, a)}$`
-					if (tex_fraction(b, a) != tex_fraction_reduite(b, a)) { texte_corr += `$=-${tex_fraction_reduite(b, a)}$` }
-					texte_corr += ' ou ' + `$x=-${tex_fraction(d, c)}$`
-					if (tex_fraction(d, c) != tex_fraction_reduite(d, c)) { texte_corr += `$=-${tex_fraction_reduite(d, c)}$` }
-					break;
-				case 6:
-					a = randint(2, 9);	//(ax+b)(cx-d)=0 	avec b/a et d/c quelconques.
-					b = randint(1, 20, [a]);
-					c = randint(2, 9, [a]);
-					d = randint(1, 20, [b, c]);
-					texte = `$(${a}x+${b})(${c}x-${d})=0$`
-					texte_corr = 'Un produit est nul si l\'un au moins de ses facteurs est nul.'
-					texte_corr += '<br>' + `$(${a}x+${b})(${c}x-${d})=0$`
-					texte_corr += '<br> Soit ' + `$${a}x+${b}=0$` + ' ou ' + `$${c}x-${d}=0$`
-					texte_corr += '<br> Donc ' + `$${a}x=${0 - b}$` + ' ou ' + `$${c}x=${d}$`
-					texte_corr += '<br> Donc ' + `$x=-${tex_fraction(b, a)}$`
-					if (tex_fraction(b, a) != tex_fraction_reduite(b, a)) { texte_corr += `$=-${tex_fraction_reduite(b, a)}$` }
-					texte_corr += ' ou ' + `$x=${tex_fraction(d, c)}$`
-					if (tex_fraction(d, c) != tex_fraction_reduite(d, c)) { texte_corr += `$=${tex_fraction_reduite(d, c)}$` }
-
-					break;
-			}
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				// alert(this.liste_questions)
-				// alert(this.liste_corrections)
-				i++;
-			}
-			cpt++;
-		}
-		liste_de_question_to_contenu(this);
-	}
-	this.besoin_formulaire_numerique = ['Niveau de difficulté', 4, '1 : Coefficient de x = 1\n 2 : Coefficient de x>1 et solutions entières\n 3 : Solutions rationnelles\n 4 : Mélange des 3 autres niveaux'];
-}
-
-/**
- * Résoudre une équation de type x²=a
-* @auteur Jean-Claude Lhote
-* 3L15
-*/
-
-function Resoudre_une_equation_x2_egal_A() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.titre = "Résoudre une équation du second degré";
-	this.consigne = "Résoudre les équations suivantes";
-	this.nb_questions = 5;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.sup = 1;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1.5
-	this.spacing = 1
-
-
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		let liste_fractions = [[1, 2], [1, 3], [2, 3], [1, 4], [3, 4], [1, 5], [2, 5], [3, 5], [4, 5],
-		[1, 6], [5, 6], [1, 7], [2, 7], [3, 7], [4, 7], [5, 7], [6, 7], [1, 8], [3, 8], [5, 8], [7, 8],
-		[1, 9], [2, 9], [4, 9], [5, 9], [7, 9], [8, 9], [1, 10], [3, 10], [7, 10], [9, 10]]
-		let liste_type_de_questions = []
-		switch (parseInt(this.sup)) {
-			case 1: liste_type_de_questions = combinaison_listes([1], this.nb_questions);
-				break;
-			case 2: liste_type_de_questions = combinaison_listes([2], this.nb_questions);
-				break;
-			case 3: liste_type_de_questions = combinaison_listes([3], this.nb_questions);
-				break;
-			case 4: liste_type_de_questions = combinaison_listes([1, 2, 3], this.nb_questions);
-
-		}
-		for (let i = 0, fraction, ns, ds, a, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-
-			switch (liste_type_de_questions[i]) {
-				case 1: a = randint(1, 20); // x²=a*a donc x=a ou -a.
-					texte = `$x^2=${a * a}$`
-					texte_corr = `$x^2=${a * a}$ équivaut à $x = \\sqrt{${a * a}}$ ou $x = -\\sqrt{${a * a}}$<br>Soit $x = ${a}$ ou $x = -${a}$<br>`
-					texte_corr += `Il est équivalent de résoudre $x^2 - ${a * a}=0$ c'est à dire $x^2 - ${a}^{2}=0$ <br>Soit $(x - ${a})(x + ${a})=0$ qui donne les deux solutions ci-dessus. `
-					break;
-				case 2: // x²=(ns*ns)/(ds*ds) solutions rationnelles
-					fraction = choice(liste_fractions);
-					ns = fraction[0]
-					ds = fraction[1]
-					texte = `$x^2=\\dfrac{${ns * ns}}{${ds * ds}}$`
-					texte_corr = `$x^2=\\dfrac{${ns * ns}}{${ds * ds}}$ équivaut à $x = \\sqrt{\\dfrac{${ns * ns}}{${ds * ds}}}$ ou $x = -\\sqrt{\\dfrac{${ns * ns}}{${ds * ds}}}$<br>Soit $x = \\dfrac{${ns}}{${ds}}$ ou $x = -\\dfrac{${ns}}{${ds}}$<br>`
-					texte_corr += `Il est équivalent de résoudre $x^2 - \\dfrac{${ns * ns}}{${ds * ds}}=0$ c'est à dire $x^2 - (\\dfrac{${ns}}{${ds}})^{2}=0$<br>Soit $(x - \\dfrac{${ns}}{${ds}})(x + \\dfrac{${ns}}{${ds}})=0$ qui donne les deux solutions ci-dessus. `
-					break;
-
-				case 3: a = randint(2, 50, [4, 9, 16, 25, 36, 49]); 	//solution irrationnelles
-					texte = `$x^2=${a}$`
-					texte_corr = `$x^2=${a}$ équivaut à $x = \\sqrt{${a}}$ ou $x = -\\sqrt{${a}}$<br>`
-					texte_corr += `Il est équivalent de résoudre $x^2 - ${a}=0$  c'est à dire $x^2 - (\\sqrt{${a}})^{2}=0$<br>Soit $(x - \\sqrt{${a}})(x + \\sqrt{${a}})=0$ qui donne les deux solutions ci-dessus. `
-					break;
-
-			}
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				// alert(this.liste_questions)
-				// alert(this.liste_corrections)
-				i++;
-			}
-			cpt++;
-		}
-		liste_de_question_to_contenu(this);
-	}
-	this.besoin_formulaire_numerique = ['Niveau de difficulté', 4, '1 : solutions entières\n 2 : solutions rationnelles\n 3 : Solutions irrationnelles\n 4 : Mélange des 3 autres niveaux'];
-}
-
-/**
- * 3F1-act - Notion de fonction - vocabulaire
- * L’objectif de revenir sur l'introduction de la notion de fonction et son vocabulaire
- * On base l'exercice sur des calculs simples de type périmètres, aires, double, triple, nombre de diviseurs
- * ATTENTION BUG SVG DONC LES ANIMATIONS SONT FILMEES A PARTIR DE CELLES GENEREES PAR LA FONCTION SVG_machine_maths() SOUS FIREFOX
- * DE FACON A AVOIR UN RENDU UNIFORME QUEL QUE SOIT LE NAVIGATEUR ON REND LES ANIMATIONS PAR DES VIDEOS
- * ON LAISSE LA PIROUETTE DE DETECTION DU USERAGENT EN COMMENTAIRE EN ATTENDANT DE TROUVER UNE SOLUTION DE RENDU LATEX DANS SVG UNIVERSELLE
- * @Auteur Sébastien Lozano
- */
-
-function fonction_notion_vocabulaire() {
-	'use strict';
-	Exercice.call(this); // Héritage de la classe Exercice()
-	this.sup = 1;
-	this.titre = "Fonctions : Notion et vocabulaire";
-	// pas de différence entre la version html et la version latex pour la consigne
-	this.consigne = `Étudier différents procédés de calcul.`;
-	sortie_html ? this.spacing = 3 : this.spacing = 1;
-	sortie_html ? this.spacing_corr = 2 : this.spacing_corr = 1;
-	//this.nb_questions;// = 4;
-	this.nb_questions_modifiable = false;
-	//this.correction_detaillee_disponible = true;
-	this.nb_cols = 1;
-	this.nb_cols_corr = 1;
-	this.sup = 5;
-	this.liste_packages = `bclogo`;
-
-	var num_ex = '3F1-act'; // pour rendre unique les id des SVG, en cas d'utilisation dans plusieurs exercices y faisant appel
-
-	if (sortie_html) {
-		var pourcentage = '100%'; // pour l'affichage des svg. On a besoin d'une variable globale
-		var hauteur_svg = 100;
-	} else { // sortie LaTeX
-
-	};
-	this.nouvelle_version = function (numero_de_l_exercice) {
-		let type_de_questions;
-
-		this.liste_questions = []; // Liste de questions
-		this.liste_corrections = []; // Liste de questions corrigées
-		this.contenu = ''; // Liste de questions
-		this.contenu_correction = ''; // Liste de questions corrigées
-		let type_de_questions_disponibles;
-		this.sup = Number(this.sup); // attention le formulaire renvoie un string, on a besoin d'un number pour le switch !
-			switch (this.sup) {
-				case 1 :
-					type_de_questions_disponibles = [1];
-					this.nb_questions = 1;
-					break;
-				case 2 :
-					type_de_questions_disponibles = [2];
-					this.nb_questions = 1;
-					break;
-				case 3 :
-					type_de_questions_disponibles = [3];
-					this.nb_questions = 1;
-					break;
-				case 4 :
-					type_de_questions_disponibles = [4];
-					this.nb_questions = 1;
-					break;
-				case 5 :
-					type_de_questions_disponibles = [1,2,3,4];
-					this.nb_questions = 4;
-					break;										
-			}
-
-		
-		//let type_de_questions_disponibles = [1];
-		let liste_type_de_questions = combinaison_listes_sans_changer_ordre(type_de_questions_disponibles, this.nb_questions);
-
-		this.introduction = lampe_message({
-			titre: `Introduction`,
-			texte: `Lorsqu'un nombre $\\textit{x}$ entre dans une machine mathématique , celle-ci renvoie à la sortie un nombre appelé $\\textit{image de x}$.<br>
-				On dit que le nombre de départ est un $\\textit{antécédent}$ du nombre qu'on trouve à la sortie.<br>
-				Ces machines sont appelées $\\textit{fonctions}$, on a l'habitude de leur donner des noms $\\textit{f}$ ou $\\textit{g}$ ou $\\textit{h} \\ldots$
-				<br>`,
-			couleur: `nombres`
-		});
-
-		if (sortie_html) { // les boutons d'aide uniquement pour la version html
-			this.bouton_aide = modal_pdf(numero_de_l_exercice, "pdf/FicheFonctions-3F1-act.pdf", "Aide mémoire sur les fonctions (Sébastien Lozano)", "Aide mémoire")
-			this.bouton_aide += modal_video('conteMathsFonctions', 'videos/Fonctions.mp4', 'Petit conte mathématique', 'Intro Vidéo');
-			this.introduction += machine_maths_video(`videos/machineMathsIntro.mp4`);
-		} else { // sortie LaTeX
-			this.introduction += tikz_machine_maths('maths', '---', `Proc\\acute{e}d\\acute{e}`, 'de\\,calcul', `ant\\acute{e}c\\acute{e}dent`, `\\textit{x}`, `image`, `\\textit{y}`);
-		};
-		for (let i = 0, x, y, z, texte, texte_corr, cpt = 0; i < this.nb_questions && cpt < 50;) {
-			type_de_questions = liste_type_de_questions[i];
-
-			if (sortie_html) {
-				var id_unique = `${num_ex}_${i}_${Date.now()}`
-				var id_du_div = `div_svg${numero_de_l_exercice}${id_unique}`;
-				var id_du_div_diag = `div_svg_diag${numero_de_l_exercice}${id_unique}`;
-				var id_du_div_corr = `div_svg_corr${numero_de_l_exercice}${id_unique}`;
-			}
-			let txt_info;
-
-			switch (type_de_questions) {
-				case 1: // périmètre d'un carré de côté x			
-					var j = 0; // pour la sous-numérotation
-					// question
-					if (sortie_html) {
-						texte = `La $\\mathbf{machine\\,f}$ renvoie le ` + katex_Popup(`périmètre`, `Rappel`, `Le périmètre d'un polygone est égal à la somme des longueurs de ses côtés`) + ` d'un carré de côté $x$`;
-					} else {
-						texte = `<br>La $\\mathbf{machine\\,f}$ renvoie le \\textbf{périmètre} \\footnote{\\textbf{Rappel :} Le périmètre d'un polygone est égal à la somme des longueurs de ses côtés} d'un carré de côté $x$`;
-					}
-					texte += `<br>`;
-					// machine						
-					x = randint(2, 99);//augmenter les possibles pour éviter les questions déjà posées?	
-					if (sortie_html) {
-						texte += machine_maths_video(`videos/machineMaths-f.mp4`);
-					} else { // sortie Latex avec Tikz
-						texte += tikz_machine_maths('f', '---', `P\\acute{e}rim\\grave{e}tre`, `d'un\\,carr\\acute{e}`, `carr\\acute{e}\\,de`, `c\\hat{o}t\\acute{e}\\,${x}\\,cm`, `P\\acute{e}rim\\grave{e}tre`, `???\\,cm`);
-					};
-					// sous question a/						
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Que renvoie la machine si le côté vaut  ${x}  cm ? Formuler la réponse `;
-						texte += katex_Popup('avec le mot image', 'Image', 'La valeur du périmètre est l\'image de la valeur du côté') + `<br>`;
-						texte_corr = num_alpha(j) + ` Si le côté vaut ${x} cm alors la machine renvoie le périmètre d'un carré de côté ${x} cm, c'est à dire $${x}+${x}+${x}+${x} = 4\\times ${x} = ${4 * x}$ cm.<br>`;
-						texte_corr += `On dit que ${4 * x} est l'image de ${x} par la fonction f.<br>`;
-						j++;//incrémente la sous question	
-					} else { //sortie LaTeX
-						texte += `\\begin{enumerate}[itemsep=1em]`;
-						texte += `\\item Que renvoie la machine si le côté vaut  ${x}  cm ? Formuler la réponse avec le mot \\textbf{image} \\footnote{\\textbf{Image :} La valeur du périmètre est l\'image de la valeur du côté}`;
-						texte_corr = `\\begin{enumerate}[itemsep=1em]`;
-						texte_corr += `\\item Si le côté vaut ${x} cm alors la machine renvoie le périmètre d'un carré de côté ${x} cm, c'est à dire $${x}+${x}+${x}+${x} = 4\\times ${x} = ${4 * x}$ cm.<br>`;
-						texte_corr += `On dit que ${4 * x} est l'image de ${x} par la fonction f.`;
-					};
-
-					// sous question b/	
-					y = randint(2, 99, [x]);//augmenter les possibles pour éviter les questions déjà posées?	
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Combien vaut le côté si la machine renvoie  ${4 * y} cm ? Formuler la réponse `;
-						texte += katex_Popup('avec le mot antécédent', 'Antécédent', 'un antécédent de la valeur d\'un périmètre est une valeur du côté qui a pour image ce périmètre') + `<br>`;
-						texte_corr += num_alpha(j) + ` Si la machine renvoie un périmètre de ${4 * y} cm alors le côté du carré vaut $${4 * y}\\div 4 = ${y}$ cm.<br>`;
-						texte_corr += `On dit que ${y} est <b>un</b> antécédent de ${4 * y} par la fonction f.<br>`;
-						j++;//incrémente la sous question
-					} else { //sortie LaTeX
-						texte += `\\item Combien vaut le côté si la machine renvoie  ${4 * y} cm ? Formuler la réponse avec le mot \\textbf{antécédent} \\footnote{\\textbf{Antécédent :} Un antécédent de la valeur d\'un périmètre est une valeur du côté qui a pour image ce périmètre}`;
-						texte_corr += `\\item Si la machine renvoie un périmètre de ${4 * y} cm alors le côté du carré vaut $${4 * y}\\div 4 = ${y}$ cm.<br>`;
-						texte_corr += `On dit que ${y} est \\textbf{un} antécédent de ${4 * y} par la fonction f.`;
-					};
-
-					// sous question c/
-					z = randint(2, 99, [x, y]);//augmenter les possibles pour éviter les questions déjà posées?						
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Quelle est l'image de ${z} par la `;
-						texte += katex_Popup('fonction', 'Vocabulaire', '<b>fonction</b> est le nom que l\'on donne à ces machines mathématiques');
-						texte += ` $\\mathbf{f}$ ? &Eacute;crire la réponse sous la forme `;
-						texte += katex_Popup('$\\mathbf{f(' + z + ')=\\ldots}$', 'Notation', '4 a pour image 16 par la fonction f peut s\'écrire <b>f(4)=16</b>') + `<br>`;
-						texte_corr += num_alpha(j) + ` L'image de ${z} par la fonction f vaut $f(${z})=4\\times ${z}=${4 * z}$.<br>`;
-						j++;//incrémente la sous question	
-					} else { // sortie LaTeX
-						texte += `\\item Quelle est l'image de ${z} par la \\textbf{fonction f} \\footnote{\\textbf{Vocabulaire :} \\textit{fonction} est le nom que l\'on donne à ces machines mathématiques}`;
-						texte += ` ? \\'{E}crire la réponse sous la forme $\\mathbf{f(${z})=\\ldots}$ \\footnote{\\textbf{Notation :} 4 a pour image 16 par la fonction f peut s\'écrire $\\mathbf{f(4)=16}$}`;
-						texte_corr += `\\item L'image de ${z} par la fonction f vaut $f(${z})=4\\times ${z}=${4 * z}$.`;
-					};
-
-					// sous question d/
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Que renvoie la machine si le côté vaut $x$ cm ?`;
-						texte += ` &Eacute;crire la réponse sous la forme `;
-						texte += katex_Popup('$\\mathbf{f(\\textbf{\\textit{x}})=\\ldots}$', 'Notation', '4 a pour image 16 par la fonction f peut s\'écrire <b>f(4)=16</b>') + `<br>`;
-						texte_corr += num_alpha(j) + ` Si le côté vaut $x$ la machine renvoie $x+x+x+x$ ce qui est équivalent à $4\\times x$ .<br>`;
-						texte_corr += ` L'image de $x$ par la fonction f vaut $4\\times x$ donc $f(x)=4\\times x$.<br>`;
-						j++;//incrémente la sous question	
-					} else { // sortie LaTeX
-						texte += `\\item   Que renvoie la machine si le côté vaut $x$ cm ?`;
-						texte_corr += `\\item  Si le côté vaut $x$ la machine renvoie $x+x+x+x$ ce qui est équivalent à $4\\times x$ .`;
-						texte += ` \\'{E}crire la réponse sous la forme $\\mathbf{f(\\textbf{\\textit{x}})=\\ldots}$ \\footnote{\\textbf{Notation :} 4 a pour image 16 par la fonction f peut s\'écrire $\\mathbf{f(4)=16}$}`;
-						texte_corr += ` L'image de $x$ par la fonction f vaut $4\\times x$ donc $f(x)=4\\times x$.`;
-					};
-
-					// sous question e/
-					txt_info = `Voici le diagramme d'une machine qui triple `;
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Comme dans l’exemple ci-dessous, écrire le diagramme de la fonction $\\mathbf{f}$.<br>`;
-						txt_info += `<div id="${id_du_div_diag}" style="width: ${pourcentage}"; height: 50px; display : table "></div>`;
-						SVG_machine_diag_3F1_act_mono(id_du_div_diag, 800, 100, 't', 'x', [['3', '3x']]);
-						texte_corr += num_alpha(j) + ` C'est une machine qui quadruple, donc sous forme de diagramme.<br>`;
-						texte_corr += `<div id="${id_du_div_corr}" style="width: ${pourcentage}"; height: 50px; display : table "></div>`;
-						SVG_machine_diag_3F1_act_mono(id_du_div_corr, 800, 100, 'f', 'x', [['4', '4x']]);
-						j++;//incrémente la sous question
-					} else { // sortie LaTeX
-						texte += `\\item   Comme dans l’exemple ci-dessous, écrire le diagramme de la fonction $\\mathbf{f}$.<br>`;
-						txt_info += '<br>' + tikz_machine_diag(`t`, `x`, [[`\\times 3`, `3x`]]);
-						texte_corr += `\\item  C'est une machine qui quadruple, donc sous forme de diagramme.<br>`;
-						texte_corr += tikz_machine_diag(`f`, `x`, [[`\\times 4`, `4x`]]);
-					};
-					texte += info_message({
-						titre: 'Exemple',
-						texte: txt_info,
-						couleur: 'nombres'
-					});
-
-					// sous question f/
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Ecrire maintenant la fonction f en utilisant la forme  `;
-						texte += katex_Popup('$\\mathbf{f:\\textbf{\\textit{x}}\\longmapsto \\ldots}$', 'Notation', '4 a pour image 16 par la fonction f peut s\'écrire $\\mathbf{f:4\\longmapsto 16}$');						
-						texte_corr += num_alpha(j) + ` L'image de $x$ par la fonction f vaut $4\\times x$ donc $f:x\\longmapsto 4\\times x$.<br>`;
-						j++;//incrémente la sous question
-					} else { // sortie LaTeX
-						texte += `\\item   Ecrire maintenant la fonction f en utilisant la forme $\\mathbf{f:\\textbf{\\textit{x}}\\longmapsto \\ldots}$ \\footnote{\\textbf{Notation :} 4 a pour image 16 par la fonction f peut s\'écrire $\\mathbf{f:4\\longmapsto 16}$},`;
-						//texte += ` écrire la réponse à la question d/`;
-						texte_corr += `\\item  L'image de $x$ par la fonction f vaut $4\\times x$ donc $f:x\\longmapsto 4\\times x$.`;
-						texte += `\\end{enumerate}`;
-						texte_corr += `\\end{enumerate}`;
-					};
-					break;
-				case 2: // aire d'un carré de côté x
-					var j = 0; // pour la sous-numérotation
-					if (sortie_html) {
-						texte = `La $\\textbf{machine\\,g}$ renvoie ` + katex_Popup('l\'aire', 'Rappel', 'L\'aire d\'un carré est égale au produit de la longueur de son côté par lui-même.') + ` d'un carré de côté $x$`;
-					} else {
-						texte = `<br>La $\\textbf{machine\\,g}$ renvoie \\textbf{l\'aire} \\footnote{\\textbf{Rappel :} L\'aire d\'un carré est égale au produit de la longueur de son côté par lui-même.} d'un carré de côté $x$`;
-					}
-					texte += `<br>`;
-					// machine
-					x = randint(2, 99);//augmenter les possibles pour éviter les questions déjà posées?	
-					if (sortie_html) {
-						texte += machine_maths_video(`videos/machineMaths-g.mp4`);
-					} else { // sortie Latex avec Tikz
-						texte += tikz_machine_maths('g', '---', `Aire`, `d'un\\,carr\\acute{e}`, `carr\\acute{e}\\,de`, `c\\hat{o}t\\acute{e}\\,${x}\\,cm`, `Aire`, `???\\,cm^2`);
-					};
-					// sous question a/	
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Que renvoie la machine si le côté vaut  ${x}  cm ? Formuler la réponse `;
-						texte += katex_Popup('avec le mot image', 'Image', 'la valeur de l\'aire est l\'image de la valeur du côté') + `<br>`;
-						texte_corr = num_alpha(j) + ` Si le côté vaut ${x} cm alors la machine renvoie l'aire d'un carré de côté ${x} cm, c'est à dire $${x}\\times ${x}=${tex_nombre(x * x)}\\,cm^2$.<br>`;
-						texte_corr += `On dit que ${nombre_avec_espace(x * x)} est l'image de ${x} par la fonction g.<br>`;
-						j++;//incrémente la sous question
-					} else { //sortie LaTeX
-						texte += `\\begin{enumerate}[itemsep=1em]`;
-						texte += `\\item  Que renvoie la machine si le côté vaut  ${x}  cm ? Formuler la réponse `;
-						texte += `avec le mot \\textbf{image} \\footnote{\\textbf{Image :} La valeur de l\'aire est l\'image de la valeur du côté.}`;
-						texte_corr = `\\begin{enumerate}[itemsep=1em]`;
-						texte_corr += `\\item Si le côté vaut ${x} cm alors la machine renvoie l'aire d'un carré de côté ${x} cm, c'est à dire $${x}\\times ${x}=${tex_nombre(x * x)}\\,cm^2$.<br>`;
-						texte_corr += `On dit que ${nombre_avec_espace(x * x)} est l'image de ${x} par la fonction g.`;
-					};
-
-					// sous question b/	
-					y = randint(2, 99, [x]);//augmenter les possibles pour éviter les questions déjà posées?	
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Combien vaut le côté si la machine renvoie  ${nombre_avec_espace(y * y)} cm<sup>2</sup> ? Formuler la réponse `;
-						texte += katex_Popup('avec le mot antécédent', 'Antécédent', 'un antécédent de la valeur d\'une aire est une valeur du côté qui a pour image cette aire') + `<br>`;
-						texte_corr += num_alpha(j) + ` Si la machine renvoie une aire de $${tex_nombre(y * y)}\\,cm^2$ alors le côté du carré vaut $\\sqrt{${tex_nombre(y * y)}}=${y}\\,cm$.<br>`;
-						texte_corr += `On dit que ${y} est <b>un</b> antécédent de ${y * y} par la fonction g.<br>`;
-						j++;//incrémente la sous question	
-					} else { //sortie LaTeX
-						texte += `\\item  Combien vaut la longueur du côté si la machine renvoie  ${nombre_avec_espace(y * y)} $cm^2$ ? Formuler la réponse `;
-						texte += `avec le mot \\textbf{antécédent} \\footnote{\\textbf{Antécédent :} Un antécédent de la valeur d\'une aire est une valeur du côté qui a pour image cette aire}`;
-						texte_corr += `\\item Si la machine renvoie une aire de $${tex_nombre(y * y)}\\,cm^2$ alors le côté du carré vaut $\\sqrt{${tex_nombre(y * y)}}=${y}\\,cm$.<br>`;
-						texte_corr += `On dit que ${y} est \\textbf{un} antécédent de ${nombre_avec_espace(y * y)} par la fonction g.`;
-					};
-
-					// sous question c/
-					z = randint(2, 99, [x, y]);//augmenter les possibles pour éviter les questions déjà posées?							
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Quelle est l'image de ${z} par la `;
-						texte += katex_Popup('fonction', 'Vocabulaire', '<b>fonction</b> est le nom que l\'on donne à ces machines mathématiques');
-						texte += ` $g$ ? &Eacute;crire la réponse sous la forme `;
-						texte += katex_Popup('$\\mathbf{g(' + z + ')=\\ldots}$', 'Notation', '4 a pour image 16 par la fonction g peut s\'écrire <b>g(4)=16</b>') + `<br>`;
-						texte_corr += num_alpha(j) + ` L'image de ${z} par la fonction g vaut $g(${z})=${z}\\times ${z}=${tex_nombre(z * z)}$.<br>`;
-						j++;//incrémente la sous question
-					} else { // sortie LaTeX
-						texte += `\\item  Quelle est l'image de ${z} par la `;
-						texte += `\\textbf{fonction g} \\footnote{\\textbf{Vocabulaire :} \\textit{fonction} est le nom que l\'on donne à ces machines mathématiques.}`;
-						texte += ` ? \\'{E}crire la réponse sous la forme `;
-						texte += `$\\mathbf{g(${z})=\\ldots}$ \\footnote{\\textbf{Notation :} 4 a pour image 16 par la fonction g peut s\'écrire \\textbf{g(4)=16}}`;
-						texte_corr += `\\item L'image de ${z} par la fonction g vaut $g(${z})=${z}\\times ${z}=${tex_nombre(z * z)}$.`;
-					};
-
-					// sous question d/
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Que renvoie la machine si le côté vaut $x$ cm ?`;
-						texte += ` &Eacute;crire la réponse sous la forme `;
-						texte += katex_Popup('$\\mathbf{g(\\textbf{\\textit{x}})=\\ldots}$', 'Notation', '4 a pour image 16 par la fonction g peut s\'écrire <b>g(4)=16</b>') + `<br>`;
-						texte_corr += num_alpha(j) + ` Si le côté vaut $x$ la machine renvoie $x\\times x$ ce qui est équivalent à $x^2$ .<br>`;
-						texte_corr += ` L'image de $x$ par la fonction g vaut $x^2$ donc $g(x)=x^2$.<br>`;
-						j++;//incrémente la sous question	
-					} else {
-						texte += `\\item  Que renvoie la machine si le côté vaut $x$ cm ?`;
-						texte_corr += `\\item Si le côté vaut $x$ la machine renvoie $x\\times x$ ce qui est équivalent à $x^2$ .`;
-						texte += ` \\'{E}crire la réponse sous la forme $\\mathbf{g(\\textbf{\\textit{x}})=\\ldots}$ \\footnote{\\textbf{Notation :} 4 a pour image 16 par la fonction g peut s\'écrire $\\mathbf{g(4)=16}$}`;
-						texte_corr += ` L'image de $x$ par la fonction g vaut $x^2$ donc $g(x)=x^2$.`;
-
-					};
-
-					// sous question e/
-					txt_info = `Voici le diagramme d'une machine qui double `;
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Comme dans l’exemple ci-dessous, écrire le diagramme de la fonction $\\mathbf{g}$.<br>`;
-						txt_info += `<div id="${id_du_div_diag}" style="width: ${pourcentage}"; height: 50px; display : table "></div>`;
-						SVG_machine_diag_3F1_act_mono(id_du_div_diag, 800, 100, 'g', 'x', [['2', '2x']]);
-						texte_corr += num_alpha(j) + ` C'est une machine qui multiplie un nombre par lui-même, donc sous forme de diagramme.<br>`;
-						texte_corr += `<div id="${id_du_div_corr}" style="width: ${pourcentage}"; height: 50px; display : table "></div>`;
-						SVG_machine_diag_3F1_act_mono(id_du_div_corr, 800, 100, 'g', 'x', [['x', 'x²']]);
-						j++;//incrémente la sous question
-					} else {
-						texte += `\\item  Comme dans l’exemple ci-dessous, écrire le diagramme de la fonction $\\mathbf{g}$.<br>`;
-						txt_info += '<br>' + tikz_machine_diag(`g`, `x`, [[`\\times 2`, `2x`]]);
-						texte_corr += `\\item C'est une machine qui multiplie un nombre par lui-même, donc sous forme de diagramme.<br>`;
-						texte_corr += tikz_machine_diag(`g`, `x`, [[`\\times x`, `x^2`]]);
-					};
-					texte += info_message({
-						titre: 'Exemple',
-						texte: txt_info,
-						couleur: 'nombres'
-					});
-
-					// sous question f/
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Ecrire maintenant la fonction g en utilisant la forme `;
-						texte += katex_Popup('$\\mathbf{g:\\textbf{\\textit{x}} \\longmapsto \\ldots}$', 'Notation', '4 a pour image 16 par la fonction g peut s\'écrire $\\mathbf{g:4\\longmapsto 16}$');						
-						texte_corr += num_alpha(j) + ` L'image de $x$ par la fonction g vaut $x\\times x=x^2$ donc $g:x\\longmapsto x\\times x=x^2$.<br>`;
-						j++;//incrémente la sous question
-					} else { // sortie LaTeX
-						texte += `\\item  Ecrire maintenant la fonction g en utilisant la forme `;
-						texte += `$\\mathbf{g:\\textbf{\\textit{x}} \\longmapsto \\ldots}$ \\footnote{\\textbf{Notation :} 4 a pour image 16 par la fonction g peut s\'écrire $\\mathbf{g:4\\longmapsto 16}$'}`;
-						texte_corr += `\\item L'image de $x$ par la fonction g vaut $x\\times x=x^2$ donc $g:x\\longmapsto x\\times x=x^2$.`;
-						texte += `\\end{enumerate}`;
-						texte_corr += `\\end{enumerate}`;
-					};
-					break;
-				case 3: // somme de 1 et du triple de x
-					var j = 0; // pour la sous-numérotation
-					// consigne
-					if (!sortie_html) {
-						texte=`<br>`;
-					} else {
-						texte=``;
-					};
-
-					texte += `La $\\mathbf{machine\\,h}$ renvoie la somme du triple du nombre de départ et de 1.`;
-					texte += `<br>`;
-					// machine
-					x = randint(2, 99);//augmenter les possibles pour éviter les questions déjà posées?	
-					if (sortie_html) {
-						texte += machine_maths_video(`videos/machineMaths-h.mp4`);
-					} else { // sortie Latex avec Tikz
-						texte += tikz_machine_maths('h', '---', `Multiplier\\,par\\,3`, `Ajouter\\,1`, `nombre\\,de`, `d\\acute{e}part\\,${x}`, `nombre\\,de`, `sortie\\,?`);
-					};
-					// sous question a/
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Que renvoie la machine si le nombre de départ vaut  ${x} ? Formuler la réponse `;
-						texte += katex_Popup('avec le mot image', 'Image', 'l\'image de la valeur à la sortie de la machine') + `<br>`;
-						texte_corr = num_alpha(j) + ` Si le nombre de départ vaut ${x} alors la machine renvoie $3\\times${x} + 1 = ${3 * x + 1}$<br>`;
-						texte_corr += `On dit que ${3 * x + 1} est l'image de ${x} par la fonction g.<br>`;
-						j++;//incrémente la sous question
-					} else { //sortie LaTeX
-						texte += `\\begin{enumerate}[itemsep=1em]`;
-						texte += `\\item  Que renvoie la machine si le nombre de départ vaut  ${x} ? Formuler la réponse `;
-						texte += `avec le mot \\textbf{image} \\footnote{\\textbf{Image :} L\'image de la valeur à la sortie de la machine.}`;
-						texte_corr = `\\begin{enumerate}[itemsep=1em]`;
-						texte_corr += `\\item Si le nombre de départ vaut ${x} alors la machine renvoie $3\\times${x} + 1 = ${3 * x + 1}$<br>`;
-						texte_corr += `On dit que ${3 * x + 1} est l'image de ${x} par la fonction g.`;
-					};
-
-					// sous question b/
-					y = randint(2, 99, [x]);//augmenter les possibles pour éviter les questions déjà posées?	
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Combien vaut le nombre de départ si la machine renvoie  ${3 * y + 1} ? Formuler la réponse `;
-						texte += katex_Popup('avec le mot antécédent', 'Antécédent', 'un antécédent d\'une valeur de sortie est une valeur du nombre de départ dont l\'image est ce nombre de sortie') + `<br>`;
-						texte_corr += num_alpha(j) + ` Si la machine renvoie $${3 * y + 1}$ alors le nombre de départ vaut $(${3 * y + 1}-1)\\div 3=${y}$<br>`;
-						texte_corr += `On dit que ${y} est <b>un</b> antécédent de ${3 * y + 1} par la fonction g.<br>`;
-						j++;//incrémente la sous question
-					} else { //sortie LaTeX
-						texte += `\\item  Combien vaut le nombre de départ si la machine renvoie  ${3 * y + 1} ? Formuler la réponse `;
-						texte += `avec le mot \\textbf{antécédent} \\footnote{\\textbf{Antécédent :} Un antécédent d\'une valeur de sortie est une valeur du nombre de départ dont l\'image est ce nombre de sortie.}`;
-						texte_corr += `\\item Si la machine renvoie $${3 * y + 1}$ alors le nombre de départ vaut $(${3 * y + 1}-1)\\div 3=${y}$<br>`;
-						texte_corr += `On dit que ${y} est \\textbf{un} antécédent de ${3 * y + 1} par la fonction g.`;
-					};
-
-					// sous question c/
-					z = randint(2, 99, [x, y]);//augmenter les possibles pour éviter les questions déjà posées?	
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Quelle est l'image de ${-z} par la `;
-						texte += katex_Popup('fonction', 'Vocabulaire', '<b>fonction</b> est le nom que l\'on donne à ces machines mathématiques');
-						texte += ` $h$ ? &Eacute;crire la réponse sous la forme `;
-						texte += katex_Popup('$\\mathbf{h(' + (-z) + ')=\\ldots}$', 'Notation', '4 a pour image 16 par la fonction h peut s\'écrire <b>h(4)=16</b>') + `<br>`;
-						texte_corr += num_alpha(j) + ` L'image de ${-z} par la fonction h vaut $h(${-z})=3\\times (${-z})+1=${-3 * z + 1}$.<br>`;
-						j++;//incrémente la sous question
-					} else { // sortie LaTeX
-						texte += `\\item  Quelle est l'image de ${-z} par la `;
-						texte += `\\textbf{fonction h} \\footnote{\\textbf{Vocabulaire :} \\textit{fonction} est le nom que l\'on donne à ces machines mathématiques}`;
-						texte += ` ? \\'{E}crire la réponse sous la forme `;
-						texte += `$\\mathbf{h(${-z})=\\ldots}$ \\footnote{\\textbf{Notation : } 4 a pour image 16 par la fonction h peut s\'écrire \\textbf{h(4)=16}}`;
-						texte_corr += `\\item L'image de ${-z} par la fonction h vaut $h(${-z})=3\\times (${-z})+1=${-3 * z + 1}$.`;
-					};
-
-					// sous question d/
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Que renvoie la machine si le côté vaut $x$ ?`;
-						texte += ` &Eacute;crire la réponse sous la forme `;
-						texte += katex_Popup('$\\mathbf{h(\\textbf{\\textit{x}})=\\ldots}$', 'Notation', '4 a pour image 16 par la fonction h peut s\'écrire <b>h(4)=16</b>') + `<br>`;
-						texte_corr += num_alpha(j) + ` Si le côté vaut $x$ la machine renvoie $3\\times x + 1$ ce qui est équivalent à $3x + 1$ .<br>`;
-						texte_corr += ` L'image de $x$ par la fonction h vaut $3\\times x + 1$ donc $f(x)=3\\times x + 1$.<br>`;
-						j++;//incrémente la sous question
-					} else { // sortie LaTeX
-						texte += `\\item  Que renvoie la machine si le côté vaut $x$ ?`;
-						texte_corr += `\\item Si le côté vaut $x$ la machine renvoie $3\\times x + 1$ ce qui est équivalent à $3x + 1$ .`;
-						texte += ` \\'{E}crire la réponse sous la forme $\\mathbf{h(\\textbf{\\textit{x}})=\\ldots}$ \\footnote{\\textbf{Notation :} 4 a pour image 16 par la fonction h peut s\'écrire $\\mathbf{h(4)=16}$}`;
-						texte_corr += ` L'image de $x$ par la fonction h vaut $3x + 1$ donc $h(x)=3x+1$.`;
-
-						j++;//incrémente la sous question
-					};
-
-					// sous question e/
-					txt_info = `Voici le diagramme d'une machine qui double puis qui ajoute 5 `;
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Comme dans l’exemple ci-dessous, écrire le diagramme de la fonction $\\mathbf{h}$.<br>`;
-						txt_info += `<div id="${id_du_div_diag}" style="width: ${pourcentage}"; height: 50px; display : table "></div>`;
-						SVG_machine_diag_3F12(id_du_div_diag, 800, 100, 'h', 'x', [['2', '2x'], ['5', '2x+5']]);
-						texte_corr += num_alpha(j) + ` C'est une machine qui triple un nombre et ajoute 1, donc sous forme de diagramme.<br>`;
-						texte_corr += `<div id="${id_du_div_corr}" style="width: ${pourcentage}"; height: 50px; display : table "></div>`;
-						SVG_machine_diag_3F12(id_du_div_corr, 800, 100, 'h', 'x', [['3', '3x'], ['1', '3x+1']]);
-						j++;//incrémente la sous question
-					} else {
-						texte += `\\item  Comme dans l’exemple ci-dessous, écrire le diagramme de la fonction $\\mathbf{h}$.<br>`;
-						txt_info += '<br>' + tikz_machine_diag(`h`, `x`, [[`\\times 2`, `2x`], [`+5`, `2x+5`]]);
-						texte_corr += `\\item C'est une machine qui triple un nombre et ajoute 1, donc sous forme de diagramme.<br>`;
-						texte_corr += tikz_machine_diag(`h`, `x`, [[`\\times 3`, `3x`], [`+1`, `3x+1`]]);
-					};
-					texte += info_message({
-						titre: 'Exemple',
-						texte: txt_info,
-						couleur: 'nombres'
-					});
-
-					// sous question f/
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Ecrire maintenant la fonction h en utilisant la forme `;
-						texte += katex_Popup('$\\mathbf{h:\\textbf{\\textit{x}} \\longmapsto \\ldots}$', 'Notation', '4 a pour image 16 par la fonction h peut s\'écrire $\\mathbf{h:4\\longmapsto16}$');						
-						texte_corr += num_alpha(j) + ` L'image de $x$ par la fonction h vaut $3\\times x +1= 3x + 1$ donc $h : x \\longmapsto 3\\times x + 1$ soit $h : x \\longmapsto 3x + 1$.<br>`;
-						j++;//incrémente la sous question
-					} else { // sortie LaTeX
-						texte += `\\item  Ecrire maintenant la fonction h en utilisant la forme `;
-						texte += `$\\mathbf{h:\\textbf{\\textit{x}} \\longmapsto \\ldots}$ \\footnote{\\textbf{Notation :} 4 a pour image 16 par la fonction h peut s\'écrire $\\mathbf{h:4\\longmapsto16}$}`;						
-						texte_corr += `\\item L'image de $x$ par la fonction h vaut $3\\times x +1= 3x + 1$ donc $h : x \\longmapsto 3\\times x + 1$ soit $h : x \\longmapsto 3x + 1$.`;
-						texte += `\\end{enumerate}`;
-						texte_corr += `\\end{enumerate}`;
-					};
-					break;
-				case 4: // nombre de diviseurs de x entier
-					var j = 0; // pour la sous-numérotation
-					// consigne
-					if (!sortie_html) {
-						texte = `<br>`;
-					} else {
-						texte=``;
-					};
-					texte += `La $\\mathbf{machine\\,d}$, qui n'accepte que des nombres entiers positifs, renvoie le nombre de diviseurs du nombre de départ.`;
-					texte += `<br>`;
-					// machine
-					x = randint(2, 51);//augmenter les possibles pour éviter les questions déjà posées?						
-					if (sortie_html) {
-						texte += machine_maths_video(`videos/machineMaths-d.mp4`);
-					} else { // sortie Latex avec Tikz
-						texte += tikz_machine_maths('d', '---', `nombre \\, total`, `de \\, diviseurs`, `nombre\\,de`, `d\\acute{e}part\\,${x}`, `nombre \\, de`, `diviseurs`);
-					};
-					// sous question a/
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Que renvoie la machine si le nombre de départ vaut  ${x} ? Formuler la réponse `;
-						texte += katex_Popup('avec le mot image', 'Image', 'l\'image de la valeur à la sortie de la machine') + `<br>`;
-						texte_corr = num_alpha(j) + ` Pour trouver la liste des diviseurs de ${x} on cherche tous les produits de deux facteurs qui donnent ${x}<br>`;
-						j++;//incrémente la sous question
-					} else { //sortie LaTeX
-						texte += `\\begin{enumerate}[itemsep=1em]`;
-						texte += `\\item Que renvoie la machine si le nombre de départ vaut  ${x} ? Formuler la réponse `;
-						texte += `avec le mot \\textbf{image} \\footnote{\\textbf{Image : } L\'image de la valeur à la sortie de la machine}`;
-						texte_corr = `\\begin{enumerate}[itemsep=1em]`;
-						texte_corr += `\\item Pour trouver la liste des diviseurs de ${x} on cherche tous les produits de deux facteurs qui donnent ${x}<br>`;
-					};
-					if (liste_diviseurs(x).length % 2 == 0) {//si il y a un nombre pair de diviseurs
-						for (let m = 0; m < (liste_diviseurs(x).length / 2); m++) {
-							texte_corr += `$` + liste_diviseurs(x)[m] + `\\times` + liste_diviseurs(x)[(liste_diviseurs(x).length - m - 1)] + `$<br>`;
-						};
-					} else {
-						for (let m = 0; m < ((liste_diviseurs(x).length - 1) / 2); m++) {
-							texte_corr += `$` + liste_diviseurs(x)[m] + `\\times` + liste_diviseurs(x)[(liste_diviseurs(x).length - m - 1)] + `$<br>`;
-						};
-						texte_corr += `$` + liste_diviseurs(x)[(liste_diviseurs(x).length - 1) / 2] + `\\times` + liste_diviseurs(x)[(liste_diviseurs(x).length - 1) / 2] + `$<br>`;
-					};
-					texte_corr += `Chacun des facteurs de la liste ci-dessus est un diviseur de ${x}<br>`;
-					texte_corr += `La liste des diviseurs de ${x} est donc ` + liste_diviseurs(x) + `; Cette liste compte ` + liste_diviseurs(x).length + ` nombres. <br>`;
-					texte_corr += `Donc ` + liste_diviseurs(x).length + ` est l'image de ${x} par la fonction d.`;
-					if (sortie_html) {
-						texte_corr += `<br>`;
-					};
-
-					// sous question b/
-					x = randint(1, 9);//augmenter les possibles pour éviter les questions déjà posées?
-					if (sortie_html) {
-						//texte += num_alpha(j) + ` Quelle est une valeur possible du nombre de départ si la machine renvoie  2 ?<br>`;
-						texte += num_alpha(j) + ` Quelle est une valeur possible du nombre de départ si la machine renvoie  2 ? En existe-t-il plusieurs ?<br>`;
-						texte_corr += num_alpha(j) + ` Si la machine renvoie 2 alors le nombre de départ  a exactement 2 diviseurs, tous les`;
-						texte_corr += katex_Popup('nombres premiers', 'Nombre premier', 'Un nombre entier est un <b>nombre premier</b> si il a exactement deux diviseurs, 1 et lui-même.');
-						texte_corr += `conviennent.<br>`;
-						texte_corr += `2 est premier donc 2 est <b>un</b> antécédent de 2 par la fonction d.<br>`;
-						texte_corr += `7 est premier donc 7 est <b>un autre</b> antécédent de 2 par la fonction d.<br>`;
-						j++;//incrémente la sous question
-					} else {
-						//texte += `\\item Quelle est une valeur possible du nombre de départ si la machine renvoie  2 ?`;
-						texte += `\\item Quelle est une valeur possible du nombre de départ si la machine renvoie  2 ? En existe-til plusieurs ?`;
-						texte_corr += ` \\item Si la machine renvoie 2 alors le nombre de départ  a exactement 2 diviseurs, tous les`;
-						texte_corr += `\\textbf{nombres premiers} \\footnote{\\textbf{Nombre premier :} Un nombre entier est un \\textbf{nombre premier} si il a exactement deux diviseurs, 1 et lui-même.}`;
-						texte_corr += `conviennent.<br>`;
-						texte_corr += `2 est premier donc 2 est \\textbf{un} antécédent de 2 par la fonction d.<br>`;
-						texte_corr += `7 est premier donc 7 est \\textbf{un autre} antécédent de 2 par la fonction d.`;
-					};
-
-					// sous question c/
-					x = randint(51, 99);//augmenter les possibles pour éviter les questions déjà posées?	
-					if (sortie_html) {
-						texte += num_alpha(j) + ` Quelle est l'image de ${x} par la `;
-						texte += katex_Popup('fonction', 'Vocabulaire', '<b>fonction</b> est le nom que l\'on donne à ces machines mathématiques');
-						texte += ` $d$ ? &Eacute;crire la réponse sous la forme `;
-						texte += katex_Popup('$\\mathbf{d(' + (x) + ')=\\ldots}$', 'Notation', '4 a pour image 16 par la fonction d peut s\'écrire <b>d(4)=16</b>') + `<br>`;
-						texte_corr += num_alpha(j) + ` Pour trouver l'image de ${x} on peut par exemple chercher tous ses diviseurs et les compter<br>`;
-						j++;//incrémente la sous question
-					} else { // sortie LaTeX
-						texte += `\\item Quelle est l'image de ${x} par la `;
-						texte += `\\textbf{fonction d} \\footnote{\\textbf{Vocabulaire :} \\textit{fonction} est le nom que l\'on donne à ces machines mathématiques.}`;
-						texte += ` ? \\'{E}crire la réponse sous la forme `;
-						texte += `$\\mathbf{d(` + (x) + `)=\\ldots}$ \\footnote{\\textbf{Notation :} 4 a pour image 16 par la fonction d peut s\'écrire \\textbf{d(4)=16}}`;
-						texte_corr += `\\item Pour trouver l'image de ${x} on peut par exemple chercher tous ses diviseurs et les compter<br>`;
-					};
-					if (liste_diviseurs(x).length % 2 == 0) {//si il y a un nombre pair de diviseurs
-						for (let m = 0; m < (liste_diviseurs(x).length / 2); m++) {
-							texte_corr += `$` + liste_diviseurs(x)[m] + `\\times` + liste_diviseurs(x)[(liste_diviseurs(x).length - m - 1)] + `$<br>`;
-						};
-					} else {
-						for (let m = 0; m < ((liste_diviseurs(x).length - 1) / 2); m++) {
-							texte_corr += `$` + liste_diviseurs(x)[m] + `\\times` + liste_diviseurs(x)[(liste_diviseurs(x).length - m - 1)] + `$<br>`;
-						};
-						texte_corr += `$` + liste_diviseurs(x)[(liste_diviseurs(x).length - 1) / 2] + `\\times` + liste_diviseurs(x)[(liste_diviseurs(x).length - 1) / 2] + `$<br>`;
-					};
-					texte_corr += `La liste des diviseurs de ${x} est donc `;
-					texte_corr += liste_diviseurs(x)[0];
-					for (let k = 1; k < liste_diviseurs(x).length; k++) {
-						texte_corr += ` ; ` + liste_diviseurs(x)[k];
-					};
-					texte_corr += ` ; Cette liste compte ` + liste_diviseurs(x).length + ` nombres.<br> `;
-					texte_corr += `Donc ` + liste_diviseurs(x).length + ` est l'image de ${x} par la fonction d.`;
-					if (sortie_html) {
-						texte_corr += `<br>`;
-					};
-
-					// sous question d/
-					if (sortie_html) {
-						// texte += num_alpha(j) + ` Peut-on trouver deux antécédents de 3 par la fonction d ?<br>`;
-						texte += num_alpha(j) + ` Peut-on trouver plusieurs antécédents de 3 par la fonction d ? Qu'ont-ils de commun ?<br>`;
-						texte_corr += num_alpha(j) + ` Il faut trouver des nombres qui ont exactement 3 diviseurs.<br>`;
-						j++;//incrémente la sous question
-					} else {
-						//texte += `\\item  Peut-on trouver deux antécédents de 3 par la fonction d ?`;
-						texte += `\\item  Peut-on trouver plusieurs antécédents de 3 par la fonction d ? Qu'ont-ils de commun ?`;
-						texte_corr += `\\item Il faut trouver des nombres qui ont exactement 3 diviseurs.<br>`;
-					}
-					texte_corr += `La liste des diviseurs de 9 est `;
-					texte_corr += liste_diviseurs(9)[0];
-					for (let k = 1; k < liste_diviseurs(9).length; k++) {
-						texte_corr += ` ; ` + liste_diviseurs(9)[k];
-					};
-					texte_corr += ` ; Cette liste compte ` + liste_diviseurs(9).length + ` nombres, `;
-					texte_corr += `donc 9 est un antécédent de 3 par la fonction d.<br>`;
-					texte_corr += `La liste des diviseurs de 25 est `;
-					texte_corr += liste_diviseurs(25)[0];
-					for (let k = 1; k < liste_diviseurs(25).length; k++) {
-						texte_corr += ` ; ` + liste_diviseurs(25)[k];
-					};
-					texte_corr += ` ; Cette liste compte ` + liste_diviseurs(25).length + ` nombres, `;
-					texte_corr += `donc 25 est un antécédent de 3 par la fonction d.<br>`;
-					texte_corr += `Tu peux en trouver d'autres, qu'ont ils de commun ?`
-					if (!sortie_html) {
-						texte += `\\end{enumerate}`;
-						texte_corr += `\\end{enumerate}`;
-					};
-					break;
-			};
-
-			if (this.liste_questions.indexOf(texte) == -1) { // Si la question n'a jamais été posée, on en créé une autre
-				this.liste_questions.push(texte);
-				this.liste_corrections.push(texte_corr);
-				i++;
-			}
-			cpt++
-		}
-
-		liste_de_question_to_contenu(this);
-	}
-	this.besoin_formulaire_numerique = ['Type de fonction',5,"1 : Périmètre d'un carré\n2 : Aire d'un carré\n3 : Somme de 1 et du triple du nombre de départ\n4 : Nombre de diviseurs d'un entier positif\n5 : Les quatre"]; 
-};
+;
 
 /**
  * 3F12 Notion de fonction - Vocabulaire
